% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{abridge}
\alias{abridge}
\alias{abridge.demog_change_component_df}
\title{Abridge a \code{ccmpp_input_df} object}
\usage{
abridge(x, ...)

\method{abridge}{demog_change_component_df}(
  x,
  span_abridged = NULL,
  age_start_abridged = NULL,
  time_start_abridged = NULL,
  out_class = class(x)[1],
  ...
)
}
\arguments{
\item{x}{An object inheriting from
\code{\link{demog_change_component_df}}.}

\item{...}{Other arguments passed to other methods.}

\item{age_start_abridged}{Vector of \emph{new} age_start values;
\code{x} will be abridged such that \code{age_start_abridged}
defines the new age groups.}

\item{out_class}{See same argument for
\code{\link{aggregate.demog_change_component_df}}.}
}
\value{
An object of the same class as \code{x}, abridged.
}
\description{
This is a generic function with methods for the various sub-classes
of \code{ccmpp_input_df}. It abridges the first argument, \code{x},
such that the result has spans or ages and times as supplied in the
\code{span_abridged} or \code{..._start_abridged} arguments. You
can only specify one of these arguments as they others are then
determined. The result must have the same value for \code{age_span}
and \code{time_span}. Abridging is done within levels of
\dQuote{sex} and \dQuote{indicators} \emph{if} \code{x} has those
dimensions.
}
\details{
This function calls
\code{\link{aggregate.demog_change_component_df}} to do the actual
abridging.

If you get an error along the lines of \dQuote{result is invalid as
a member of the class} or similar, it means that, after abridging,
the result is no longer valid as a member of the class \code{x}
started as. This would occur, for example, if \code{x} inherits
from \code{ccmpp_input_df} and you only supplied either
\dQuote{age} or \dQuote{time} arguments. Try supplying both
\dQuote{age} and \dQuote{time} arguments, e.g., \code{... age_span
= 5, time_span = 5}.
}
\seealso{
\code{\link[stats]{aggregate}} which this function relies
    on, \code{link{collapse_demog_dimension}} for a related
    function.
}
\author{
Mark Wheldon
}
