% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{as.data.frame.vital_rate_df_time_age_sex}
\alias{as.data.frame.vital_rate_df_time_age_sex}
\title{Coerce a \code{vital_rate_df_time_age_sex} to \code{data.frame}}
\usage{
\method{as.data.frame}{vital_rate_df_time_age_sex}(x)
}
\arguments{
\item{x}{An object of class \code{vital_rate_df_time_age_sex}.}
}
\value{
A \code{data.frame}.
}
\description{
The class-specific attributes will be lost.
}
\author{
Mark Wheldon
}
