% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_ccmpp_input_list.R
\name{as.list.ccmpp_input_list}
\alias{as.list.ccmpp_input_list}
\title{Coerce a \code{ccmpp_input_list} to a plain list.}
\usage{
\method{as.list}{ccmpp_input_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ccmpp_input_list}.}
}
\value{
A list with special attributes dropped.
}
\description{
The result will have \code{\link{class}} \code{list} and
class-specific attributes, such as \dQuote{age_span}, will be
dropped. Neither the classes nor the class-specific attributes of
the component data frames, all of which  inherit from
\code{\link{ccmpp_inut_df}}, will be dropped (see
\code{\link{as.data.frame.ccmpp_input_df}} if you want to do that).
}
\seealso{
\code{\link{as.data.frame.ccmpp_input_df}} to drop classes
    and attributes of component data frames,
    \code{\link{ccmpp_input_list}} to create objects of class \code{ccmpp_input_list}
}
\author{
Mark Wheldon
}
