% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_vital_rate_df.R
\name{as.list.vital_rate_df}
\alias{as.list.vital_rate_df}
\title{Coerce a \code{vital_rate_df} to \code{list}}
\usage{
\method{as.list}{vital_rate_df}(x)
}
\arguments{
\item{x}{An object of class \code{vital_rate_df}.}
}
\value{
A \code{list}.
}
\description{
The class-specific attributes will be lost.
}
\author{
Mark Wheldon
}
