% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{as.numeric.demog_change_component_df}
\alias{as.numeric.demog_change_component_df}
\title{Coerce a \code{demog_change_component_df} to \code{data.frame}}
\usage{
\method{as.numeric}{demog_change_component_df}(x)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}
}
\value{
A \code{data.frame}.
}
\description{
The class-specific attributes will be lost.
}
\author{
Mark Wheldon
}
