% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{as.numeric.vital_rate_df}
\alias{as.numeric.vital_rate_df}
\title{Coerce a \code{vital_rate_df} to \code{data.frame}}
\usage{
\method{as.numeric}{vital_rate_df}(x)
}
\arguments{
\item{x}{An object of class \code{vital_rate_df}.}
}
\value{
A \code{data.frame}.
}
\description{
The class-specific attributes will be lost.
}
\author{
Mark Wheldon
}
