% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{as_age_time_matrix}
\alias{as_age_time_matrix}
\alias{as_age_time_matrix.demog_change_component_df}
\title{Coerce a \code{demog_change_component_df} to \code{matrix}}
\usage{
as_age_time_matrix(x)

\method{as_age_time_matrix}{demog_change_component_df}(x)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}
}
\value{
A \code{list}.
}
\description{
An age * time matrix will be created from the input object,
\code{x}, which must not have a \dQuote{sex} dimension. To create a
matrix from such an object, first remove the \dQuote{sex} dimension
with \code{\link{subset_sex}}.
}
\author{
Mark Wheldon
}
