% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_utility_functions.R
\name{as_age_time_matrix_list}
\alias{as_age_time_matrix_list}
\alias{as_age_time_matrix_list.ccmpp_input_list}
\title{Coerce a \code{ccmpp_input_list} to a list of age-time matrices}
\usage{
as_age_time_matrix_list(x, drop_zero_fert_ages = FALSE, ...)

\method{as_age_time_matrix_list}{ccmpp_input_list}(x, drop_zero_fert_ages = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from class
\code{demog_change_component_df}.}

\item{drop_zero_fert_ages}{Logical; should rows corresponding to
\dQuote{non_zero_fert_ages} be dropped when converting
\code{fert_rate_age_f} objects.}

\item{...}{Further arguments passed to and from methods.}
}
\value{
A \code{list}.
}
\description{
A list of age * time matrices will be created from \code{x} by
calling \code{\link{as_age_time_matrix}} on each element.
}
\seealso{
\code{\link{as_age_time_matrix}}
}
\author{
Mark Wheldon
}
