% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_input_file.R
\name{basepop_adjust_1950_population}
\alias{basepop_adjust_1950_population}
\title{Adjust base population children for consistency with fertility and mortality}
\usage{
basepop_adjust_1950_population(pop_count_age_sex_base, input_file_path)
}
\arguments{
\item{input_file_path}{character string pointing to excel input file}
}
\value{
data frame with adjusted 1950 base population
}
\description{
Invokes DemoTools::basepop_five to adjust counts of children under age 10 to be consistent with 1950
fertility and mortality
}
\details{
uses life tables, asfr and srb from the excel input file
}
\author{
Sara Hertog
}
