% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_wpp_1950_2020_ccmpp_inputs_1x1}
\alias{canada_wpp_1950_2020_ccmpp_inputs_1x1}
\title{Example inputs file for 1x1 cohort component population projection for Canada 1950-2020}
\format{
A list with elements
\describe{
\item{LocID}{A numeric location identifier for the country, in this case 124".}
\item{variant}{A character string identifying the projection variant, in this case "Estimates".}
\item{year_base}{The base (starting) year of the projection, in this case 1950.}
\item{year_stop}{The year to stop the projection, in this case 2020.}
\item{age}{A vector of ages associated with the input data, in this case 0:100.}
\item{age_width}{The width of age groups associated with the input data, in this case 1.}
\item{pop_count_age_sex_base}{A data frame with population counts by age and sex in the base year.}
\item{life_table_age_sex}{A data frame of life table values by single year of age and sex by year from 1950 to 2019.}
\item{fert_rate_age_f}{A data frame of age-specific fertility rates by single year of age by year from 1950 to 2019.}
\item{mig_net_count_age_sex}{A data frame of net migrant counts by single year of age and sex by year from 1950-2019.}
\item{srb}{A data frame of the sex ratio at birth (males/females) by year from 1950-2019.}
\item{mig_assumption}{A data frame with indicator of whether migration should be accounted at "end" of period or "even" over period by year from 1950-2019.}
}
}
\source{
Interpolated from estimates published in United Nations, \emph{World Population Prospects 2019}.
}
\usage{
canada_wpp_1950_2020_ccmpp_inputs_1x1
}
\description{
All inputs needed to perform a cohort component population projection for Canada
by single year of age and in one-year steps for the projection (1x1) from 1950-2020
}
\keyword{datasets}
