% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_wpp_1950_2020_ccmppWPP_outputs}
\alias{canada_wpp_1950_2020_ccmppWPP_outputs}
\title{Example outputs file for 1x1 cohort component population projection for Canada 1950-2020}
\format{
A list with elements
\describe{
\item{LocID}{A numeric location identifier for the country, in this case 124.}
\item{variant}{A character string identifying the projection variant, in this case "Estimates".}
\item{pop_count_age_sex_1x1}{A data frame with population counts in the value column, by year, sex and single year of age.}
\item{pop_count_age_sex_5x1}{A data frame with population counts in the value column, by year, sex and five year age group.}
\item{pop_count_tot_sex}{A data frame with total population counts in the value column, by year and sex.}
\item{birth_count_age_b_1x1}{A data frame with birth counts (both sexes) in the value column, by year and single year of age of mother.}
\item{birth_count_age_b_5x1}{A data frame with birth counts (both sexes) in the value column, by year and five year age group of mother.}
\item{fert_rate_age_f_1x1}{A data frame with females' age-specific fertility rates in the value column, by year and single year of age.}
\item{fert_rate_age_f_5x1}{A data frame with females' age-specific fertility rates in the value column, by year and five year age group.}
\item{fert_pct_age_f_1x1}{A data frame with females' percentage age-specific fertility rates in the value column, by year and single year of age.}
\item{fert_pct_age_f_5x1}{A data frame with females' percentage age-specific fertility rates in the value column, by year and five year age group.}
\item{fert_rate_tot_f}{A data frame with females' total fertility rates in the value column, by year.}
\item{fert_mean_age_f}{A data frame with females' mean age at childbearing in the value column, by year.}
\item{srb}{A data frame with the sex ratio at birth (M/F) in the value column, by year. }
\item{death_count_age_sex_1x1}{A data frame with death counts in the value column, by year, sex and single year of age.}
\item{death_count_age_sex_5x1}{A data frame with death counts in the value column, by year, sex and five year age group.}
\item{death_count_tot_sex}{A data frame with total death counts in the value column, by year and sex.}
\item{death_count_cohort_sex_1x1}{A data frame with death counts in the value column, by year, sex and single year age cohort.}
\item{exposure_count_age_sex_1x1}{A data frame with person-years of exposure in the value column, by year, sex and single year of age.}
\item{exposure_count_age_sex_5x1}{A data frame with person-years of exposure in the value column, by year, sex and five year age group.}
\item{lt_complete_age_sex}{A data frame with life table column labels in the indicator column and life table values
in the value column, by year, sex and single year of age.}
\item{lt_abridged_age_sex}{A data frame with life table column labels in the indicator column and life table values
in the value column, by year, sex and abridged age groups (0,1,5,10,15....)}
\item{lt_summary}{A data frame with life table summary indicator labels in the indicator column and life table values
in the value column, by sex (example labels: lt_1q0, lt_4q1, lt_35q15, lt_e0, lt_e60, etc.).}
\item{mig_net_count_age_sex_1x1}{A data frame with net migrant counts in the value column, by year, sex and single year of age.}
\item{mig_net_count_age_sex_5x1}{A data frame with net migrant counts in the value column, by year, sex and five year age group.}
\item{mig_net_count_tot_sex}{A data frame with total net migrant counts in the value column, by year and sex.}
\item{mig_assumption}{A data frame with whether migration is accounted at "end" of period or "even" over period in the value column, by year.}
}
}
\usage{
canada_wpp_1950_2020_ccmppWPP_outputs
}
\description{
All estimates generated for one country-variant by the full cohort component population projection.
This example  for Canada by single year of age and in one-year steps (1x1) from 1950-2020
Output is from function call
project_ccmppWPP_one_country_variant(ccmpp_input = canada_wpp_1950_2020_ccmpp_inputs_1x1)
}
\keyword{datasets}
