% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df.R
\name{ccmpp_input_df}
\alias{ccmpp_input_df}
\title{Constructor for class \code{ccmpp_input_df}}
\usage{
ccmpp_input_df(
  x,
  dimensions = attr(x, "dimensions"),
  age_span = attr(x, "age_span"),
  time_span = attr(x, "time_span"),
  value_type = attr(x, "value_type"),
  ...
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{age_span}{Scalar indicating the span of the age groups. If
\code{NULL} and a column \dQuote{age_span} is present in
\code{x} an attempt will be made to infer it from that column.}

\item{time_span}{Scalar indicating the span of the time periods;
similar behaviour to \code{age_span}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{ccmpp_input_df}.
}
\description{
**TBC** More strict version of \code{\link{demog_change_component_df}}.
}
\author{
Mark Wheldon
#' @export
}
