% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{ccmpp_input_list}
\alias{ccmpp_input_list}
\title{Constructor for class \code{ccmpp_input_list}}
\usage{
ccmpp_input_list(
  pop_count_base_input_df,
  life_table_input_df,
  fert_rate_input_df,
  srb_input_df,
  mig_net_count_input_df,
  mig_net_rate_input_df,
  mig_net_count_tot_input_df,
  mig_parameter_input_df
)
}
\value{
An object of class \code{ccmpp_input_list}.
}
\description{
**TBC**
}
\seealso{
Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{fert_rate_input_df}()},
\code{\link{life_table_input_df}()},
\code{\link{mig_net_count_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{mig_parameter_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{srb_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
