% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_output_df.R
\name{ccmpp_output_df}
\alias{ccmpp_output_df}
\title{Constructor for class \code{ccmpp_output_df}}
\usage{
ccmpp_output_df(
  x,
  dimensions = attr(x, "dimensions"),
  value_type = attr(x, "value_type"),
  value_scale = attr(x, "value_scale"),
  ...
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{ccmpp_output_df}.
}
\description{
**TBC** More strict version of \code{\link{demog_change_component_df}}.
}
\seealso{
Other ccmpp_output_objects: 
\code{\link{coerce_ccmpp_output_df}},
\code{\link{coerce_pop_count_age_sex}},
\code{\link{pop_count_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_output_objects}
