% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_aggregate.R
\name{ccmppWPP_aggregate}
\alias{ccmppWPP_aggregate}
\title{aggregate ccmppWPP outputs across multiple location ids}
\usage{
ccmppWPP_aggregate(
  locationIDs,
  base_year,
  last_year,
  intermediate_output_folder,
  LocIDnew,
  LocNamenew
)
}
\arguments{
\item{locationIDs}{xx}
}
\value{
xx
}
\description{
xx
}
\details{
xx
}
\author{
Sara Hertog
}
