% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_input_file.R
\name{ccmppWPP_input_file_estimates}
\alias{ccmppWPP_input_file_estimates}
\title{read ccmppWPP inputs from country-specific excel input files}
\usage{
ccmppWPP_input_file_estimates(input_file_path)
}
\arguments{
\item{input_file_path}{character. string pointing to the file path for the Excel input file for one country.}
}
\value{
a list of data frames -- one for each type of input needed to produce the estimates variant with the ccmpp.
attributes assigned to the list describe the location id and name, a0 rule and variant
}
\description{
This function reads the ccmppWPP inputs from the country specific excel files.  Specifically,
it reads from parameters, pop_count_age_sex_base, fert_rate_age_f, srb, mort_params, life_table_age_sex,
mig_net_count_age_sex, mig_net_count_tot_b, and mig_parameters sheets of the country input files.  It selects only
years bounded by the estimation period.
}
\author{
Sara Hertog
}
