% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_workflow_one_country_variant.R
\name{ccmppWPP_workflow_one_country_variant}
\alias{ccmppWPP_workflow_one_country_variant}
\title{Full ccmppWPP implementation for one country and variant}
\usage{
ccmppWPP_workflow_one_country_variant(wpp_input)
}
\arguments{
\item{wpp_input}{list of input objects required for one country-variant}
}
\value{
a list of objects including projected population, deaths, exposures, life tables, births, fertility
rates and proportions, and net migrant counts
}
\description{
This function takes the wpp input list of objects for a given country and variant
and walks through the entire ccmppWPP workflow to produce all of the outputs published in the WPP.
}
\details{
This function will accept any width of age group and projection horizon as long as those are equal
e.g., use z=1 for a 1x1 projection of population by single year of age and 1 year projection horizon
or use z=5 for a 5x5 projection of population by 5-year age groups and 5-year projection horizon
}
\author{
Sara Hertog
}
