% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_input_file.R
\name{census_back_project_1950}
\alias{census_back_project_1950}
\title{Adjust base population children for consistency with fertility and mortality}
\usage{
census_back_project_1950(
  census_protocol_adjusted,
  census_reference_date,
  life_table_age_sex
)
}
\arguments{
\item{census_protocol_adjusted}{data frame with census population by single year of age and sex output from census adjustment protocol}

\item{census_reference_date}{numeric. census reference date as decimal}

\item{life_table_age_sex}{data frame. the life_table_age_sex table from input file}
}
\value{
data frame with 1950 base population back projected from the earliest census
}
\description{
Back projects from census one year at a time using input survival ratios, then interpolates to 1 Jan 1950
}
\details{
extends both population and life tables to oag = 130+; NO INTERNATIONAL MIGRATION
}
\author{
Sara Hertog
}
