% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_years}
\alias{census_years}
\title{Years censuses were conducted by country}
\format{
\describe{
  \item{}{}
  \item{}{}
  ...
}
}
\source{
\url{}
}
\usage{
census_years
}
\description{
A \code{\link[tibble]{tibble}} containing the years in which census
data are available for most countries and areas.
}
\keyword{datasets}
