% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df_utility_functions.R
\name{change_ccmpp_input_df_span}
\alias{change_ccmpp_input_df_span}
\alias{span<-}
\alias{span<-.ccmpp_input_df}
\title{Change the span of a \code{ccmpp_input_df} object}
\usage{
span(x, ...) <- value

\method{span}{ccmpp_input_df}(x, ...) <- value
}
\arguments{
\item{x}{An object inheriting from \code{\link{ccmpp_input_df}}.}

\item{value}{A numeric scalar for the new span.}

\item{...}{Passed to or from other methods.}
}
\value{
\code{x} with spans set to \code{value}.
}
\description{
This is a generic function with methods for the various sub-classes
of \code{ccmpp_input_df}. Valid objects inheriting from
\code{ccmpp_input_df} have a common value for \dQuote{age_span} and
\dQuote{time_span}. This function provides a way of coherently
changing this value, ensuring that the result is still a valid
member of its class.
}
\author{
Mark Wheldon
}
