% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df.R
\name{coerce_ccmpp_input_df}
\alias{coerce_ccmpp_input_df}
\alias{as_ccmpp_input_df}
\alias{as_ccmpp_input_df.default}
\alias{as_ccmpp_input_df.data.frame}
\alias{as_ccmpp_input_df.matrix}
\alias{as_ccmpp_input_df.ccmpp_input_df}
\alias{is_ccmpp_input_df}
\title{Coerce to a \code{ccmpp_input_df}}
\usage{
as_ccmpp_input_df(x, ...)

\method{as_ccmpp_input_df}{default}(x, ...)

\method{as_ccmpp_input_df}{data.frame}(
  x,
  value_type = attr(x, "value_type"),
  value_scale = attr(x, "value_scale"),
  ...
)

\method{as_ccmpp_input_df}{matrix}(
  x,
  value_type = attr(x, "value_type"),
  value_scale = attr(x, "value_scale"),
  ...
)

\method{as_ccmpp_input_df}{ccmpp_input_df}(x, ...)

is_ccmpp_input_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{ccmpp_input_df} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
