% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{coerce_ccmpp_input_list}
\alias{coerce_ccmpp_input_list}
\alias{as_ccmpp_input_list}
\alias{as_ccmpp_input_list.default}
\alias{as_ccmpp_input_list.list}
\alias{as_ccmpp_input_list.ccmpp_input_list}
\alias{is_ccmpp_input_list}
\title{Coerce to a \code{ccmpp_input_list}}
\usage{
as_ccmpp_input_list(x, ...)

\method{as_ccmpp_input_list}{default}(x, ...)

\method{as_ccmpp_input_list}{list}(x, ...)

\method{as_ccmpp_input_list}{ccmpp_input_list}(x, ...)

is_ccmpp_input_list(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{ccmpp_input_list} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}} for an important note on validation.
}
\author{
Mark Wheldon
}
