% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_output_df.R
\name{coerce_ccmpp_output_df}
\alias{coerce_ccmpp_output_df}
\alias{as_ccmpp_output_df}
\alias{as_ccmpp_output_df.default}
\alias{as_ccmpp_output_df.data.frame}
\alias{as_ccmpp_output_df.matrix}
\alias{as_ccmpp_output_df.ccmpp_output_df}
\alias{is_ccmpp_output_df}
\title{Coerce to a \code{ccmpp_output_df}}
\usage{
as_ccmpp_output_df(x, ...)

\method{as_ccmpp_output_df}{default}(x, ...)

\method{as_ccmpp_output_df}{data.frame}(x, ...)

\method{as_ccmpp_output_df}{matrix}(x, ...)

\method{as_ccmpp_output_df}{ccmpp_output_df}(x, ...)

is_ccmpp_output_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{ccmpp_output_df} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}

Other ccmpp_output_objects: 
\code{\link{ccmpp_output_df}()},
\code{\link{pop_count_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_output_objects}
