% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{coerce_demog_change_component_df}
\alias{coerce_demog_change_component_df}
\alias{as_demog_change_component_df}
\alias{as_demog_change_component_df.data.frame}
\alias{as_demog_change_component_df.matrix}
\alias{is_demog_change_component_df}
\title{Coerce to a \code{demog_change_component_df}}
\usage{
as_demog_change_component_df(x, ...)

\method{as_demog_change_component_df}{data.frame}(x, ...)

\method{as_demog_change_component_df}{matrix}(x, ...)

is_demog_change_component_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{demog_change_component_df} if possible, or check if it is
one.
}
\section{Note on Validation}{

The method for \code{demog_change_component_df} (i.e.,
\code{as_demog_change_component_df.demog_change_component_df}) does
\emph{not} check for validity. If \code{x} is somehow an invalid
\code{demog_change_component_df} object, applying
\code{as_demog_change_component_df} will simply return it (with
\code{attr(x, "class")} potentially modifed), not signal an
error. Use \code{demog_change_component_df} on the object to ensure
validity is checked.
}

\author{
Mark Wheldon
}
