% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_fert_rate_df.R
\name{coerce_fert_rate_age_f}
\alias{coerce_fert_rate_age_f}
\alias{as_fert_rate_age_f}
\alias{as_fert_rate_age_f.default}
\alias{as_fert_rate_age_f.data.frame}
\alias{as_fert_rate_age_f.matrix}
\alias{as_fert_rate_age_f.fert_rate_age_f}
\alias{is_fert_rate_age_f}
\title{Coerce to a \code{fert_rate_age_f}}
\usage{
as_fert_rate_age_f(x, ...)

\method{as_fert_rate_age_f}{default}(x, ...)

\method{as_fert_rate_age_f}{data.frame}(x, ...)

\method{as_fert_rate_age_f}{matrix}(x, ...)

\method{as_fert_rate_age_f}{fert_rate_age_f}(x, ...)

is_fert_rate_age_f(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{fert_rate_age_f} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
