% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_net_count_tot_df.R
\name{coerce_mig_net_count_tot_b}
\alias{coerce_mig_net_count_tot_b}
\alias{as_mig_net_count_tot_b}
\alias{as_mig_net_count_tot_b.default}
\alias{as_mig_net_count_tot_b.data.frame}
\alias{as_mig_net_count_tot_b.matrix}
\alias{as_mig_net_count_tot_b.mig_net_count_tot_b}
\alias{is_mig_net_count_tot_b}
\title{Coerce to a \code{mig_net_count_tot_b}}
\usage{
as_mig_net_count_tot_b(x, ...)

\method{as_mig_net_count_tot_b}{default}(x, ...)

\method{as_mig_net_count_tot_b}{data.frame}(x, ...)

\method{as_mig_net_count_tot_b}{matrix}(x, ...)

\method{as_mig_net_count_tot_b}{mig_net_count_tot_b}(x, ...)

is_mig_net_count_tot_b(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{mig_net_count_tot_b} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
