% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_parameter_df.R
\name{coerce_mig_parameter}
\alias{coerce_mig_parameter}
\alias{as_mig_parameter}
\alias{as_mig_parameter.default}
\alias{as_mig_parameter.data.frame}
\alias{as_mig_parameter.matrix}
\alias{as_mig_parameter.mig_parameter}
\alias{is_mig_parameter}
\title{Coerce to a \code{mig_parameter}}
\usage{
as_mig_parameter(x, ...)

\method{as_mig_parameter}{default}(x, ...)

\method{as_mig_parameter}{data.frame}(x, ...)

\method{as_mig_parameter}{matrix}(x, ...)

\method{as_mig_parameter}{mig_parameter}(x, ...)

is_mig_parameter(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{mig_parameter} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
