% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mort_rate_df.R
\name{coerce_mort_rate_age_sex}
\alias{coerce_mort_rate_age_sex}
\alias{as_mort_rate_age_sex}
\alias{as_mort_rate_age_sex.default}
\alias{as_mort_rate_age_sex.data.frame}
\alias{as_mort_rate_age_sex.matrix}
\alias{as_mort_rate_age_sex.mort_rate_age_sex}
\alias{is_mort_rate_age_sex}
\title{Coerce to a \code{mort_rate_age_sex}}
\usage{
as_mort_rate_age_sex(x, ...)

\method{as_mort_rate_age_sex}{default}(x, ...)

\method{as_mort_rate_age_sex}{data.frame}(x, ...)

\method{as_mort_rate_age_sex}{matrix}(x, ...)

\method{as_mort_rate_age_sex}{mort_rate_age_sex}(x, ...)

is_mort_rate_age_sex(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{mort_rate_age_sex} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}} for an important note on validation.
}
\author{
Mark Wheldon
}
