% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_pop_count_base_df.R
\name{coerce_pop_count_age_sex_base}
\alias{coerce_pop_count_age_sex_base}
\alias{as_pop_count_age_sex_base}
\alias{as_pop_count_age_sex_base.data.frame}
\alias{as_pop_count_age_sex_base.matrix}
\alias{is_pop_count_age_sex_base}
\title{Coerce to a \code{pop_count_age_sex_base}}
\usage{
as_pop_count_age_sex_base(x, ...)

\method{as_pop_count_age_sex_base}{data.frame}(x, ...)

\method{as_pop_count_age_sex_base}{matrix}(x, ...)

is_pop_count_age_sex_base(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{pop_count_age_sex_base} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
