% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_pop_count_base_input.R
\name{coerce_pop_count_base_input_df}
\alias{coerce_pop_count_base_input_df}
\alias{as_pop_count_base_input_df}
\alias{as_pop_count_base_input_df.data.frame}
\alias{as_pop_count_base_input_df.matrix}
\alias{is_pop_count_base_input_df}
\title{Coerce to a \code{pop_count_base_input_df}}
\usage{
as_pop_count_base_input_df(x, ...)

\method{as_pop_count_base_input_df}{data.frame}(x, ...)

\method{as_pop_count_base_input_df}{matrix}(x, ...)

is_pop_count_base_input_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{pop_count_base_input_df} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}} for an important note on validation.
}
\author{
Mark Wheldon
}
