% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_srb_df.R
\name{coerce_srb}
\alias{coerce_srb}
\alias{as_srb}
\alias{as_srb.default}
\alias{as_srb.data.frame}
\alias{as_srb.matrix}
\alias{as_srb.srb}
\alias{is_srb}
\title{Coerce to a \code{srb}}
\usage{
as_srb(x, ...)

\method{as_srb}{default}(x, ...)

\method{as_srb}{data.frame}(x, ...)

\method{as_srb}{matrix}(x, ...)

\method{as_srb}{srb}(x, ...)

is_srb(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{srb} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}
}
\author{
Mark Wheldon
}
