% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_survival_ratio_input_df.R
\name{coerce_survival_ratio_input_df}
\alias{coerce_survival_ratio_input_df}
\alias{as_survival_ratio_input_df}
\alias{as_survival_ratio_input_df.default}
\alias{as_survival_ratio_input_df.data.frame}
\alias{as_survival_ratio_input_df.matrix}
\alias{as_survival_ratio_input_df.survival_ratio_input_df}
\alias{is_survival_ratio_input_df}
\title{Coerce to a \code{survival_ratio_input_df}}
\usage{
as_survival_ratio_input_df(x, ...)

\method{as_survival_ratio_input_df}{default}(x, ...)

\method{as_survival_ratio_input_df}{data.frame}(x, ...)

\method{as_survival_ratio_input_df}{matrix}(x, ...)

\method{as_survival_ratio_input_df}{survival_ratio_input_df}(x, ...)

is_survival_ratio_input_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{survival_ratio_input_df} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}} for an important note on validation.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_input_df}},
\code{\link{coerce_life_table_input_df}},
\code{\link{coerce_mig_net_count_input_df}},
\code{\link{coerce_mig_net_count_tot_input_df}},
\code{\link{coerce_mig_net_rate_input_df}},
\code{\link{coerce_mig_parameter_input_df}},
\code{\link{coerce_pop_count_base_input_df}},
\code{\link{coerce_srb_input_df}},
\code{\link{fert_rate_input_df}()},
\code{\link{life_table_input_df}()},
\code{\link{mig_net_count_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{mig_parameter_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{srb_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
