% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{col_row_names}
\alias{col_row_names}
\alias{names<-}
\alias{names<-.default}
\alias{names<-.demog_change_component_df}
\alias{rownames<-}
\alias{rownames<-.default}
\alias{rownames<-.demog_change_component_df}
\alias{colnames<-}
\alias{colnames<-.default}
\alias{colnames<-.demog_change_component_df}
\alias{dimnames<-}
\alias{dimnames<-.default}
\alias{dimnames<-.demog_change_component_df}
\title{Change column and row names of a \code{demog_change_component_df}}
\usage{
names(x) <- value

\method{names}{default}(x) <- value

\method{names}{demog_change_component_df}(x) <- value

rownames(x) <- value

\method{rownames}{default}(x) <- value

\method{rownames}{demog_change_component_df}(x) <- value

colnames(x) <- value

\method{colnames}{default}(x) <- value

\method{colnames}{demog_change_component_df}(x) <- value

dimnames(x) <- value

\method{dimnames}{default}(x) <- value

\method{dimnames}{demog_change_component_df}(x) <- value
}
\arguments{
\item{x}{a matrix-like \R object, with at least two dimensions for
    \code{colnames}.}

\item{value}{a valid value for that component of
    \code{\link[base]{dimnames}(x)}.  For a matrix or array this is either
    \code{NULL} or a character vector of non-zero length equal to the
    appropriate dimension.}
}
\value{
A \code{data.frame},
    one-dimensional vector, or scalar depending on the dimensions
    of the extracted values.
}
\description{
Changing row or columns names of \code{demog_change_component_df}s
may result in invalid members of the class. Therefore, the changed
object will no longer inherit from class
\code{demog_change_component_df} (or subclasses) and the attributes
specific to those classes will be lost.
}
\seealso{
\code{\link{demog_change_component_df}} for class
    definitions.
}
\author{
Mark Wheldon
}
