% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_vital_rate_df.R
\name{construct_vital_rate_df}
\alias{construct_vital_rate_df}
\alias{vital_rate_df}
\alias{as.vital_rate_df}
\alias{is.vital_rate_df}
\title{Constructor for class \code{vital_rate_df}}
\usage{
vital_rate_df(x, age_span = NULL, time_span = NULL, dimensions = NULL, ...)

as.vital_rate_df(x, ...)

is.vital_rate_df(x)
}
\arguments{
\item{x}{A data frame with columns \dQuote{age_start},
\dQuote{age_span}, \dQuote{sex}, \dQuote{time_start},
\dQuote{time_span}, and \dQuote{value}.}

\item{age_span}{Scalar indicating the span of the age groups. If
\code{NULL} and a column \dQuote{age_span} is present in
\code{x} an attempt will be made to infer it from that column.}

\item{time_span}{Scalar indicating the span of the time periods;
similar behaviour to \code{age_span}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{vital_rate_df}.
}
\description{

}
\details{
The class is based on a data frame with age-sex-specific vital rates for
one or more time periods in a column named \dQuote{value}.

See the help files for the other constructor functions under
\dQuote{See also} for further details specific to each type of vital rate
or parameter.
}
\section{Class Definition}{


**TBC** (details of the class)
1. Must be sorted properly.

\subsection{Attributes of objects with this class}

**TBC**

**ISSUES OUTSTANDING**
======================

SORTING:
--------

Correct ordering by 'age_start' and 'time_start' is required for
the ccmpp functions (right?) What about by 'sex'? The sample
inputs are sorted by sex first (varying slowest) but with 'male'
first, so in reverse alphabetical order.

FACTORS vs CHARACHTERS:
-----------------------

'sex' is a categorical variable and encoding it as a factor could
be justified. However it may be easier to encode as character and
have methods convert to factor when advantageous.
}

\author{
Mark Wheldon
}
\concept{vital_rate_df constructor functions}
