% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parse_ccmpp_input.R
\name{data_parse_ccmpp_input}
\alias{data_parse_ccmpp_input}
\title{Parse the objects from ccmpp_input file needed by the project_ccmpp_loop_over_time function}
\usage{
data_parse_ccmpp_input(indata)
}
\arguments{
\item{indata}{name of the standard wpp_input file}
}
\value{
a list of data.frames including base population by time and age and sex, survival ratios by time and age and sex,
fertility rates by time and age of mother, the sex ratio at birth by time, net migration counts by time and age and sex,
net migration rates by time and age and sex, total net migrant counts by time, migration parameters that control treatment
of net migration in the ccmpp
}
\description{
This function takes a wpp_input file and parses the objects needed to perform the 
full cohort-component population projection over time. Returns the list of inputs required for 
project_ccmpp_loop_over_time function
}
\author{
Sara Hertog
}
