% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reshape_wpp_output_long.R
\name{data_reshape_wpp_output_long}
\alias{data_reshape_wpp_output_long}
\title{Reshape ccmppWPP output into a single long data frame with all indicators}
\usage{
data_reshape_wpp_output_long(wpp_output)
}
\arguments{
\item{wpp_output}{list of output indicators generated by ccmppWPP_workflow_one_country_variant()}
}
\value{
a data frame with revision, variant, indicator name, time, sex, age and value
}
\description{
this function takes the list output by the cmmppWPP_workflow_one_country_variant and reshapes
it into a single long data frame
}
\details{
NA
}
\author{
Sara Hertog
}
