% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1.R
\name{DDextract_ccmppWPPinputs_tier1}
\alias{DDextract_ccmppWPPinputs_tier1}
\title{Extract data needed for probabilistic population reconstruction with ccmppWPP}
\usage{
DDextract_ccmppWPPinputs_tier1(
  LocID,
  times = 1950:2020,
  time_reference_base = "01/01/1950",
  data_source_base = "HMD",
  revision = "test",
  variant = "estimates"
)
}
\arguments{
\item{LocID}{numeric. Location id of country.}

\item{times}{numeric. Vector of years for which data should be extracted}

\item{time_reference_base}{character. Reference date of the base year for population.}

\item{data_source_base}{character. DataSourceShortName of the data source for base year population.}

\item{revision}{character. Revision attribute for output ccmppWPP_inputs list.}

\item{variant}{character. Variant attribute for output ccmppWPP_inputs list.}
}
\value{
a list of two objects. 1) the ccmppWPP_inputs list required to run the population projection; and
2) a pop_counts_age_sex_reference data frame for reconciliation
}
\description{
This function extracts from the DemoData SQL database population counts by
age and sex, annual births by sex to compute sex ratio at birth, 1x1 age specific 
fertility rates from the Human Fertility Database and mortality rates from the Human 
Mortality Database.  It compiles those data into the input file required by ccmppWPP.  It 
also compiles a refernce dataset of population counts by age and sex for various reference
periods
}
\details{
currently extracts from paperspace development server.
}
\author{
Sara Hertog
}
