% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{DDextract_get_locid}
\alias{DDextract_get_locid}
\title{Get the location ID for a DDextract}
\usage{
DDextract_get_locid(x)
}
\arguments{
\item{x}{A list of the form returned by
\code{\link{DDextract_ccmppWPPinputs_tier1}}}
}
\value{
A numeric or character location ID.
}
\description{
This will identify the location to which the data in \code{x}
pertain. Note that this will simply extract the stored value which
(currently) could be a numeric or character identifier.
}
\author{
Mark Wheldon
}
