% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{DDextract_get_pop_count_age_sex_reference}
\alias{DDextract_get_pop_count_age_sex_reference}
\title{Get reference population counts from a DDSQL query}
\usage{
DDextract_get_pop_count_age_sex_reference(
  x,
  times = c("all", "excl_baseline", "census", "census_excl_baseline")
)
}
\arguments{
\item{x}{A list of the form returned by
\code{\link{DDextract_ccmppWPPinputs_tier1}}.}

\item{times}{Either a numeric vector containing years for which
reference population counts should be extracted from \code{x},
or a character string specifying a pre-defined set of times;
see \dQuote{Details}.}

\item{...}{Passed to other methods.}
}
\value{
An object of class \code{\link{demog_change_component_df}}
    containing the reference population counts.
}
\description{
Extracts and (optionally) subsets reference population data from a
list of the kind returned by
\code{\link{DDextract_ccmppWPPinputs_tier1}}. The result is
returned as a \code{\link{demog_change_component_df}} object.
}
\details{
\code{x} must be a
\code{\link{list}} with a \code{\link{data.frame}} element named
\code{pop_count_age_sex_reference}. This \code{data.frame} must
have colums \code{time_start}, \code{time_span}, \code{age_start},
\code{age_span}, \code{sex}, \code{value}.

Argument \code{times} can be used to return counts for specific
times only. These can be specified explicitly as a numeric vector
or one of the key words \describe{
\item{\dQuote{excl_baseline}}{All years available except the
baseline year.}
\item{\dQuote{all}}{All years available.}
\item{\dQuote{census}}{Only the years censuses were
conducted. Census years are taken from the dataset
\code{"census_years"} provided with the package.}
\item{\dQuote{census_excl_baseline}}{Census years excluding the
baseline year (if the baseline year was not a census year this will
give the same result as \dQuote{census}).}}
}
\section{Note}{

Census times cannot be determined by default because the object
returned by \code{\link{DDextract_ccmppWPPinputs_tier1}} does not
record the location to which the data pertain. Use
\code{\link{get_census_years}} to separately look up census years.
}

\examples{

data("france_wpp_1950_2020_population_data")
data("census_years")

all_years <-
  DDextract_get_pop_count_age_sex_reference(france_wpp_1950_2020_population_data)

census_years_only <-
  DDextract_get_pop_count_age_sex_reference(france_wpp_1950_2020_population_data,
                                   times = "census")

selected_years <-
  DDextract_get_pop_count_age_sex_reference(france_wpp_1950_2020_population_data,
                                   times = 1950:1955)

}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
