% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{DDextract_get_pop_count_reference_times}
\alias{DDextract_get_pop_count_reference_times}
\alias{DDextract_exclude_baseline_pop_count_times}
\title{Get reference population times from DDSQL extract}
\usage{
DDextract_get_pop_count_reference_times(x)

DDextract_exclude_baseline_pop_count_times(x, times)
}
\arguments{
\item{x}{A list of the form returned by \code{\link{DDextract_ccmppWPPinputs_tier1}}.}
}
\value{
A numeric vector of times.
}
\description{
\code{DDextract_exclude_baseline_pop_count_times} subsets its argument
\code{times} by excluding the time of the baseline population
counts in the \code{ccmppWPP_inputs} element of
\code{x}. \code{DDextract_get_pop_count_reference_times} is a wrapper that
sets \code{times} to the times of the
\code{pop_count_age_sex_reference} element of \code{x}. This
function is primarily intended to be called via
\code{\link{DDextract_to_pop_count_age_sex_reference}}. See that function's
help file for more details.
}
\seealso{
\code{\link{ccmpp_input_list}}, \code{\link{pop_count_age_sex_base}}

Other DDSQL_extract: 
\code{\link{DDextract_get_pop_count_age_sex_reference}()},
\code{\link{get_census_years}()},
\code{\link{parse_pop_count_reference_times_ranges}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
