% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{DDextract_to_ccmpp_input_list}
\alias{DDextract_to_ccmpp_input_list}
\title{Convert Demo Data extract to a CCMPP input list}
\usage{
DDextract_to_ccmpp_input_list(x)
}
\arguments{
\item{x}{A list of the form returned by \code{\link{DDextract_ccmppWPPinputs_tier1}}.}
}
\value{
An object inheriting from \code{\link{ccmpp_input_list}}.
}
\description{
Takes the output of \code{\link{DDextract_ccmppWPPinputs_tier1}}
and returns an object inheriting from
\code{\link{ccmpp_input_list}}. Note that this will discard some
components of the input (e.g., the reference population counts).
}
\details{
This function extracts the CCMPP input list component from \code{x}
and checks that the components are consistent with each
other. Currenlty, this involves selecting rows for times that all
components have in common.
}
\author{
Mark Wheldon
}
