% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_helper_functions.R
\name{death_age_sex_loop_over_time}
\alias{death_age_sex_loop_over_time}
\title{Loop over time to compute age-period deaths from cohort deaths and lx and nLx life table values}
\usage{
death_age_sex_loop_over_time(dth_cohort, lx, nLx)
}
\arguments{
\item{dth_cohort}{data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the cohort-specific death counts output from the project_ccmpp_loop_over_time function}

\item{lx}{data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the lx values from the wpp inputs life table data frame}

\item{nLx}{data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the nLx values from the wpp inputs life table data frame}
}
\value{
a data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the sex- and age-specific death counts
}
\description{
Loops over time to implement the death_cohort_period_to_age_period function for multiple
periods of time.
}
\author{
Sara Hertog
}
