% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_helper_functions.R
\name{death_cohort_period_to_age_period}
\alias{death_cohort_period_to_age_period}
\title{Compute period deaths by age from cohort deaths and life table values}
\usage{
death_cohort_period_to_age_period(
  death_cohort_period,
  lx_age_period,
  nLx_age_period
)
}
\arguments{
\item{death_cohort_period}{numeric. vector of deaths by cohort computed by project_ccmpp_z_by_z.}

\item{lx_age_period}{numeric. vector life table lx values}

\item{nLx_age_period}{numeric. vector of life table nLx values}
}
\value{
a numeric vector period deaths by age.
}
\description{
This function computes the age-specific period deaths from cohort-specific deaths
and life table values lx and nLx. Approach uses cohort separation factors as described by 
Thomas Buettner in http://abcabacus.org/?page_id=599
}
\author{
Sara Hertog
}
