% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_death_probability_df.R
\name{death_probability_age_sex}
\alias{death_probability_age_sex}
\alias{death_probability_age_sex.data.frame}
\alias{death_probability_age_sex.life_table_age_sex}
\alias{death_probability_age_sex.ccmpp_input_list}
\alias{death_probability_age_sex<-}
\title{Constructor for class \code{death_probability_age_sex}}
\usage{
death_probability_age_sex(x, ...)

\method{death_probability_age_sex}{data.frame}(x)

\method{death_probability_age_sex}{life_table_age_sex}(x)

\method{death_probability_age_sex}{ccmpp_input_list}(x)

death_probability_age_sex(x) <- value
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Details}).}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{death_probability_age_sex}.
}
\description{
\code{death_probability_age_sex} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{proportion}.}
  \item{Within year and sex, age must start at 0.}}
}
\details{
Methods are defined for \code{\link{data.frame}}s and
\code{\link{ccmpp_input_list}}s, and possibly other objects as
well. The \code{data.frame} method \dQuote{constructs} an object
from \code{x}. The \code{ccmpp_input_list} method \dQuote{extracts}
an object from \code{x}. There is also a replacement function which
complements the extraction methods.
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_count_age_sex}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mortality_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{mortality_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
