% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_df_access_attr.R
\name{demog_change_component_attributes}
\alias{demog_change_component_attributes}
\alias{demog_change_component_attributes.demog_change_component_df}
\title{Extract attributes specific to \code{demog_change_component_df}s}
\usage{
demog_change_component_attributes(x)

\method{demog_change_component_attributes}{demog_change_component_df}(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
A list of attributes.
}
\description{
All attributes that are not attributes of \code{data.frame} objects
are extracted. This function is designed to work with objects of
class \code{demog_change_component_df}; behaviour for other classes is
not defined.
}
\author{
Mark Wheldon
}
