% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_df_access_attr.R
\name{demog_change_component_dimensions}
\alias{demog_change_component_dimensions}
\title{Extract the \dQuote{dimensions} attribute}
\usage{
demog_change_component_dimensions(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
The extracted attribute
}
\description{
Extracts the non-data.frame attributes. The only method that exists
is for \code{demog_change_component_df} objects.
}
\author{
Mark Wheldon
}
