% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_df_access_attr.R
\name{demog_change_component_dims}
\alias{demog_change_component_dims}
\alias{demog_change_component_dims.demog_change_component_df}
\title{Report the demographic \dQuote{dimensions}}
\usage{
demog_change_component_dims(x)

\method{demog_change_component_dims}{demog_change_component_df}(x)
}
\arguments{
\item{x}{An object for which a method exists, e.g., one inheriting
from \code{demog_change_component_df}.}
}
\value{
A character vector with demographic dimensions present.
}
\description{
Reports the demographic dimensions, in a character vector.
}
\author{
Mark Wheldon
}
