% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_fert_rate_df.R
\name{drop_zero_fert_ages}
\alias{drop_zero_fert_ages}
\alias{drop_zero_fert_ages.fert_rate_age_f}
\title{Drop zero fertility rate ages}
\usage{
drop_zero_fert_ages(x, ...)

\method{drop_zero_fert_ages}{fert_rate_age_f}(x)
}
\arguments{
\item{x}{An object inheriting from \code{fert_rate_age_df}}

\item{...}{Not implemented}
}
\value{
A \code{demog_change_component_df} with the only the rows
    corresponding to \code{non_zero_fert_ages(x)}.
}
\description{
Drops zero fertility rate ages from a \code{fert_rate_age_df}
object. Note that the result is \emph{not valid} as a member of
this class so a \code{\link{demog_change_component_df}} will be
returned.
}
\author{
Mark Wheldon
}
