% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_helper_functions.R
\name{exposure_age}
\alias{exposure_age}
\title{Approximate age-specific person-years of exposure to fertility/mortality in the period}
\usage{
exposure_age(death_age_period, nmx_age_period)
}
\arguments{
\item{death_age_period}{numeric. vector of age-period deaths}

\item{nmx_age_period}{numeric. vector of life table nMx values}
}
\value{
a numeric vector of person-years of exposure by single year of age.
}
\description{
This function takes cohort-period deaths and age-specific nMx to approximate age-specific person years of exposure
Useful for aggregating life tables across time, age, geographies
}
\author{
Sara Hertog
}
