% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_age_period.R
\name{exposure_age_period}
\alias{exposure_age_period}
\title{Compute age-specific person-years of exposure to fertility/mortality in the period}
\usage{
exposure_age_period(
  pop_age_begin,
  pop_age_end,
  migration_assumption = c("even", "end"),
  mig_net_count = 0
)
}
\arguments{
\item{pop_age_begin}{numeric. vector of population by single year of age at the beginning of the year.}

\item{pop_age_end}{numeric. vector of population by single year of age at the end of the year.}

\item{migration_assumption}{character. field indicating whether migration is treated as "end" of year or "even" over year.}

\item{mig_net_count}{numeric. vector of net migrants by single year of age between begin year and end year.}
}
\value{
a numeric vector of person-years of exposure by single year of age.
}
\description{
This function takes beginning and end period population by age and computes person-years of exposure, 
removing migrants from exposure if migration assuption is "end" of period and adding half of migrants to beginning 
of perod if migration assumption is "even" over period.
}
\details{
For "even" migration assumption, half of net migrants are added to beginning of year population for computing exposures;
for "end" migration assumption, net migrants are subtracted from end of year population for computing exposures.
}
\author{
Sara Hertog
}
