% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_helper_functions.R
\name{exposure_age_sex_loop_over_time}
\alias{exposure_age_sex_loop_over_time}
\title{Loop over time to approximate age-specific person years of exposure}
\usage{
exposure_age_sex_loop_over_time(dth_age, nmx)
}
\arguments{
\item{dth_age}{data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the sex- and age-specific death counts.}

\item{nmx}{data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the nMx values from the wpp inputs life table data frame}
}
\value{
a data frame with columns time_start, time_span, sex, age_start, age_span and value.  Value contains
the sex- and age-specific person years of exposure counts
}
\description{
Loops over time to implement the exposure_age function for multiple
periods of time.
}
\author{
Sara Hertog
}
