% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_access_attr.R
\name{extract_ccmpp_input_list_attributes}
\alias{extract_ccmpp_input_list_attributes}
\alias{age_span.ccmpp_input_list}
\alias{time_span.ccmpp_input_list}
\title{Extract attributes of a \code{ccmpp_input_list}}
\usage{
\method{age_span}{ccmpp_input_list}(x)

\method{time_span}{ccmpp_input_list}(x)
}
\arguments{
\item{x}{An object inheriting from \code{ccmpp_input_list}}
}
\value{
The requested attribute
}
\description{
These are convenience functions for accessing specific attributes
of objects inheriting from \code{ccmpp_input_list}.
}
\seealso{
Other extract_attributes: 
\code{\link{age_span}()}
}
\author{
Mark Wheldon
}
\concept{extract_attributes}
