% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_helper_functions_wpp.R
\name{fert_gross}
\alias{fert_gross}
\title{Compute gross reproductive rate from age-specific fertility rates and sex ratio at birth}
\usage{
fert_gross(fert_data_age, srb, byvar)
}
\arguments{
\item{fert_data_age}{data frame. columns "age_start" and "age_span" describe age group of mother
and "value" is age-specific fertility rates in births per woman.}

\item{srb}{data frame. column "value" is the sex ratio at birth}

\item{byvar}{character. string of column names in fert_data_age and srb by which GRR will be computed.}
}
\value{
a data frame with gross reproductive rate in the "value" field, computed by byvar columns.
}
\description{
This function computes the gross reproductive rate from the age-specific fertility rates
and sex ratio at birth of a given period.
}
\details{
accomodates any width of age interval (age_span) so long as it is constant over age.
}
\author{
Sara Hertog
}
