% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_helper_functions_wpp.R
\name{fert_mac}
\alias{fert_mac}
\title{Compute mean age at childbearing (mac) from age-specific fertility rates}
\usage{
fert_mac(fert_data_age, byvar)
}
\arguments{
\item{fert_data_age}{data frame. column "age" is age group of mother and "value" is age-specific fertility rates in births per woman.}

\item{byvar}{character. string of column names in fert_data_age by which pasfr will be computed.}
}
\value{
a data frame with mean age at childbearing in the "value" field, computed by byvar columns and age.
}
\description{
This function computes the mean age at childbearing from the age-specific fertility rates 
of a given period.
}
\details{
age/period can be 1x1 or 5x5.
}
\author{
Sara Hertog
}
