% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_helper_functions_wpp.R
\name{fert_net}
\alias{fert_net}
\title{Compute net reproductive rate from age-specific fertility rates, sex ratio at birth and female life table nLx}
\usage{
fert_net(fert_data_age, srb, nLx, byvar)
}
\arguments{
\item{fert_data_age}{data frame. columns "age_start" and "age_span" describe age group of mother
and "value" is age-specific fertility rates in births per woman.}

\item{srb}{data frame. column "value" is the sex ratio at birth}

\item{nLx}{data frame. columns "age_start" and "age_span" describe age group of mother and "value"
is the nLx column of the period life table for females}

\item{byvar}{character. string of column names in fert_data_age, srb and nLx by which NRR will be computed.}
}
\value{
a data frame with net reproductive rate in the "value" field, computed by byvar columns.
}
\description{
This function computes the net reproductive rate from the age-specific fertility rates
and sex ratio at birth and female life table nLx of a given period.
}
\details{
accomodates any width of age interval (age_span) so long as it is constant over age.
Assumes radix (l0) = 100000.
}
\author{
Sara Hertog
}
