% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_helper_functions_wpp.R
\name{fert_pasfr}
\alias{fert_pasfr}
\title{fertility computation helper functions for WPP}
\usage{
fert_pasfr(fert_data_age, byvar)
}
\arguments{
\item{fert_data_age}{data frame. columns "age_start" and "age_span" describe age group of mother
and "value" is age-specific fertility rates in births per woman.}

\item{byvar}{character. string of column names in fert_data_age by which pasfr will be computed.}
}
\value{
a data frame with percentage age specific fertility rates in the "value" field, computed by byvar
columns and age_start, age_span. Sums to 100 for each combination of byvar.
}
\description{
This function computes the percentage age-specific fertility rates from the age-specific fertility
rates of a given period.
}
\details{
Compute percentage age-specific fertility rates from age-specific fertility rates

accomodates any width of age interval (age_span) so long as it is constant over age.
}
\author{
Sara Hertog
}
