% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{fert_rate_age_f}
\alias{fert_rate_age_f}
\title{Constructor for class \code{fert_rate_age_f}}
\usage{
fert_rate_age_f(
  x,
  non_zero_fert_ages = attr(x, "non_zero_fert_ages"),
  value_scale = attr(x, "value_scale")
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{non_zero_fert_ages}{Numeric vector of unique ages indicating
the reproductive age range. See Section \dQuote{Non-zero fert ages}.}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{fert_rate_age_f}.
}
\description{
\code{fert_rate_age_f} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes three additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{There can be no 'sex' dimension; all fertility rate inputs
  must be female fertility rates.}
  \item{The additional attribute \code{non_zero_fert_ages}
  represents the reproductive age range as a vector of
  \code{age_start} values.}}
An attempt will be made to guess \code{non_zero_fert_ages} if not supplied.
}
\section{Non-zero fert ages}{

Elements of the \code{value} column corresponding to ages not
listed in \code{non_zero_fert_ages} will be overwritten with zeros
(specifically \code{as.double(0)}s).


\code{non_zero_fert_ages} must be a subset of \code{ages(x)} and
equally spaced, with spacing equal to \code{age_span(x)}.
}

\seealso{
\code{\link{validate_ccmpp_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
