% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{fert_rate_input_df}
\alias{fert_rate_input_df}
\title{Constructor for class \code{fert_rate_input_df}}
\usage{
fert_rate_input_df(
  x,
  dimensions = attr(x, "dimensions"),
  age_span = attr(x, "age_span"),
  time_span = attr(x, "time_span"),
  non_zero_fert_ages = attr(x, "non_zero_fert_ages"),
  ...
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{age_span}{Scalar indicating the span of the age groups. If
\code{NULL} and a column \dQuote{age_span} is present in
\code{x} an attempt will be made to infer it from that column.}

\item{time_span}{Scalar indicating the span of the time periods;
similar behaviour to \code{age_span}.}

\item{non_zero_fert_ages}{Numeric vector of unique ages indicating
the reproductive age range.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{fert_rate_input_df}.
}
\description{
\code{fert_rate_input_df} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{The additional attribute \code{non_zero_fert_ages}
  represents the reproductive age range as a vector of
  \code{age_start} values.}}
An attempt will be made to guess \code{non_zero_fert_ages} if not supplied.
}
\section{Non-zero fert ages}{

\code{non_zero_fert_ages} must be a subset of \code{ages(x)} and
equally spaced, with spacing equal to \code{age_span(x)}.
}

\author{
Mark Wheldon
}
