% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_fert_rate_df.R
\name{fert_rate_tot_f}
\alias{fert_rate_tot_f}
\alias{tfr}
\alias{fert_rate_tot_f.fert_rate_age_f}
\alias{fert_rate_tot_f.ccmpp_input_list}
\title{Calculate total fertility rates}
\usage{
fert_rate_tot_f(x, ...)

tfr(x, ...)

\method{fert_rate_tot_f}{fert_rate_age_f}(x)

\method{fert_rate_tot_f}{ccmpp_input_list}(x)
}
\arguments{
\item{x}{An object with information on fertility; typically
inheriting from \code{\link{fert_rate_age_f}} or
\code{\link{ccmpp_input_list}}.}

\item{...}{Passed to specific methods}
}
\value{
An object of class \code{\link{demog_change_component_df}}
}
\description{
This generic function returns total fertility rates from an object
with information on fertility.
}
\details{
\code{tfr} is an alias for this function.
}
\author{
Mark Wheldon
}
