% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france_wpp_1950_2020_population_data}
\alias{france_wpp_1950_2020_population_data}
\title{Demographic data for France, 1950--2020}
\format{
\describe{
  \item{}{}
  \item{}{}
  ...
}
}
\source{
\url{}
}
\usage{
france_wpp_1950_2020_population_data
}
\description{
A dataset containing fertility, mortality, migration and population
data for France, 1950--2000.
}
\keyword{datasets}
