% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{get_census_years}
\alias{get_census_years}
\alias{get_census_years.list}
\alias{get_census_years.numeric}
\alias{get_census_years.character}
\title{Get years in which censuses were conducted}
\usage{
get_census_years(x)

\method{get_census_years}{list}(x)

\method{get_census_years}{numeric}(x)

\method{get_census_years}{character}(x)
}
\arguments{
\item{x}{A numeric location code or location name as a character
string, or a list of the form returned by
\code{\link{DDextract_ccmppWPPinputs_tier1}}.}
}
\value{
A \emph{character} vector of years.
}
\description{
This function takes a three digit numeric code (\cite{United
Nations, 2021}) for a location (country or area) and returns the
years in which censuses were conducted. This is a convenience
function for extracting the years from
\code{data("census_years")}. This function is called by
\code{\link{DDextract_to_pop_count_age_sex_reference}}. See that
function's help file for additional notes.
}
\details{
Years in which censuses were conducted are provided in
\code{data("census_years")} for many countries. Ranges of the form
\code{"xxxx-yyyy"} (e.g., \code{"2000-2005"}) are possible if a
rolling population count was conducted over several years. France,
for example, has conducted rolling censuses.
}
\references{
United Nations (2021). Standard country or area codes for
    statistical use (M49). Department of Economic and Social
    Affairs, Statistics
    Division. https://unstats.un.org/unsd/methodology/m49/
    (retrieved 1 April 2021).
}
\seealso{
Other DDSQL_census_years: 
\code{\link{parse_census_years_ranges}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_census_years}
