% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData_extract_post_process.R
\name{get_pop_count_age_sex_reference}
\alias{get_pop_count_age_sex_reference}
\title{Make reference population counts}
\usage{
get_pop_count_age_sex_reference(x, times = NULL)
}
\arguments{
\item{x}{A list of the form returned by
\code{\link{DDextract_ccmppWPPinputs_tier1}}.}

\item{times}{Vector containing years for which reference population
counts should be extracted from \code{x}. Can be numeric or
character; see \dQuote{Details}. If \code{NULL} (default), all
times will be retained.}

\item{...}{Passed to other methods.}
}
\value{
An object of class
    \code{\link[ccmppWPP]{demog_change_component_df}} containing
    the reference population counts.
}
\description{
Extracts and (optionally) subsets reference population data from a
list returned by \code{\link{DDextract_ccmppWPPinputs_tier1}}, or
data frames extracted from such a list. If \code{x} is a
\code{list} it must have a \code{data.frame} element named
\code{\code{pop_count_age_sex_reference}}. The \code{data.frame}
method will be called on this element. For the \code{data.frame}
method, \code{x} must have colums \code{time_start},
\code{time_span}, \code{age_start}, \code{age_span}, \code{sex},
\code{value}.
}
\details{
\code{times} can be used to return counts for specific times only,
e.g., specific years. If \code{times} is \code{NULL} (default) it
will be initialized to the times of
\code{x$pop_count_age_sex_reference}. You might, instead, want to
set it to something determined by
\code{\link{get_census_years}}. Regardless,
\code{\link{extract_pop_count_reference_times}} will be called on
\code{times}. This will first discard any elements of \code{times}
not in \code{times(x$pop_count_age_sex_reference)}. Then it will
parse the result using
\code{\link{parse_pop_count_reference_times_ranges}} so that any
ranges of the form \code{"xxxx-yyyy"} are correctly converted to
numeric values. Once \code{times} is set and parsed, the time of
the baseline population counts in
\code{pop_count_base_component(x$ccmppWPP_inputs)} will be removed.

Years in which censuses were conducted are provided in
\code{data("census_years")} for many countries. Ranges of the form
\code{"xxxx-yyyy"} (e.g., \code{"2000-2005"}) are possible if a
rolling population count was conducted over several years. France,
for example, has conducted rolling censuses. In such cases, the
midpoint of the range will be calculated and trucated down to the
nearest integer.
}
\examples{

data("france_wpp_1950_2020_population_data")
data("census_years")

all_years <-
  get_pop_count_age_sex_reference(france_wpp_1950_2020_population_data)

census_years_only <-
  get_pop_count_age_sex_reference(france_wpp_1950_2020_population_data,
                                   times = get_census_years(250))

}
\seealso{
Other DDSQL_extract: 
\code{\link{get_census_years}()},
\code{\link{get_pop_count_reference_times}()},
\code{\link{parse_pop_count_reference_times_ranges}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
