% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_helper_functions_wpp.R
\name{lt_abridged_from_complete}
\alias{lt_abridged_from_complete}
\title{Compute an abridged life table from a complete life table}
\usage{
lt_abridged_from_complete(age_complete, lx_complete, sex)
}
\arguments{
\item{age_complete}{numeric. a vector of single-year ages from the complete life table.}

\item{lx_complete}{numeric. a vector of lx (life table survivors) values from the complete life table.}

\item{sex}{character. sex associated to the life table, can be "female", "male", "both", or "f", "m", "b"}
}
\value{
a long data frame with columns "indicator" containing the life table column label, "age" containing
the starting age of abridged life table age groups, and "value" containing the numeric value associated with 
the life table indicator
}
\description{
This function computes an abridged life table from the lx column of a complete life table (single 
year of age).
}
\author{
Sara Hertog
}
