% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_helper_functions_wpp.R
\name{lt_long}
\alias{lt_long}
\title{life table helper functions for WPP}
\usage{
lt_long(life_table_df)
}
\arguments{
\item{life_table_df}{data frame. life table output by the lt_abridged or lt_single_mx functions of DemoTools.}
}
\value{
a long data frame with columns "indicator" containing the life table column label, "age", and "value"
containing the numeric value associated with the life table indicator
}
\description{
This function reshapes the wide life table data frame returned by the DemoTools functions 
lt_abridged() or lt_single_mx() and retuns a long data frame with the "indicator", "age", "value" columns
used for WPP.
}
\details{
#' Take life tables output from lt functions in DemoTools and reshape to long data frame
}
\author{
Sara Hertog
}
