% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_helper_functions_wpp.R
\name{lt_summary}
\alias{lt_summary}
\title{Compute summary life table values}
\usage{
lt_summary(lt_data, byvar)
}
\arguments{
\item{lt_data}{data frame. column "indicator" is the label for the life table column, age" is age group 
and "value" is is the value of the given life table indicator.}

\item{byvar}{character. string of column names in lt_data by which the life table values are computed.}
}
\value{
a data frame with "indicator" labeling the life table summary measure and "value" containing the value of 
that measure, computed by byvar columns.
}
\description{
This function computes selected summary life table values from the output of DemoTools life table
function lt_abridged() or lt_single_mx() that has been transformed to a long dataset with the lt_long() function.
}
\details{
age/period can be 1x1 or 5x5.
}
\author{
Sara Hertog
}
