% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{make_value_product}
\alias{make_value_product}
\title{Calculate the product of values in two data frames}
\usage{
make_value_product(x, y, by_vars_names = NULL, all.x = TRUE, all.y = FALSE)
}
\arguments{
\item{x, y}{Data frames with columns \dQuote{\code{by_vars_names}}
and \dQuote{value}.}

\item{by_vars_names}{Character vector of names of columns to merge
by. By default, all columns except \dQuote{value}. See also
\dQuote{Details}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to
    the output, one for each row in \code{x} that has no matching row in
    \code{y}.  These rows will have \code{NA}s in those columns that are
    usually filled with values from \code{y}.  The default is
    \code{FALSE}, so that only rows with data from both \code{x} and
    \code{y} are included in the output.}

\item{all.y}{logical; analogous to \code{all.x}.}
}
\value{
A data frame with columns \dQuote{\code{by_vars_names}} and
    \dQuote{value}, where the latter holds the product of
    \dQuote{value}s in \code{x} and \code{y}.
}
\description{
The product of the \dQuote{value} columns in data frames \code{x}
and \code{y} is calculated and returned. \code{x} and
\code{y} are first merged on columns named \code{by_vars_names}
and these columns are in the output, which is also a data frame.
}
\details{
When \code{by_vars_names} is \code{NULL}, if either \code{x} or
\code{y} has column \code{time_span} and all entries are 0,
\dQuote{time_span} will be omitted from
\code{by_vars_names}.
}
\author{
Mark Wheldon
}
