% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{make_value_ratio}
\alias{make_value_ratio}
\title{Calculate the ratio of values in two data frames}
\usage{
make_value_ratio(num, denom, by_vars_names = NULL, all.x = TRUE, all.y = FALSE)
}
\arguments{
\item{num}{A data frame with columns \dQuote{\code{by_vars_names}}
and \dQuote{value}, holding the numerator values.}

\item{denom}{Similar to \code{num} but holding the denominator
values.}

\item{by_vars_names}{Character vector of names of columns to merge
by. By default, all columns except \dQuote{value}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to
    the output, one for each row in \code{x} that has no matching row in
    \code{y}.  These rows will have \code{NA}s in those columns that are
    usually filled with values from \code{y}.  The default is
    \code{FALSE}, so that only rows with data from both \code{x} and
    \code{y} are included in the output.}

\item{all.y}{logical; analogous to \code{all.x}.}
}
\value{
A data frame with columns \dQuote{\code{by_vars_names}} and
    \dQuote{value}, where the latter holds the ratio of
    \dQuote{value}s in \code{num} and \code{denom}.
}
\description{
The ratio of the \dQuote{value} columns in the numerator
(\code{num}) and denominator (\code{denom}) data frames is
calculated and returned. \code{num} and \code{denom} are first
merged on columns named \code{by_vars_names} and these columns are
in the output, which is also a data frame.
}
\author{
Mark Wheldon
}
