% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_helper_functions.R
\name{median_age}
\alias{median_age}
\title{Compute median age of the population from 1x1 population counts}
\usage{
median_age(pop_count_age)
}
\arguments{
\item{pop_count_age}{data frame. columns "time_start, "age_start" and "value" where value is population counts}
}
\value{
a data frame with time_start and median age of the population (years) in the "value" field
}
\description{
This function computes the median age of the population from 1x1 population counts
}
\details{
single year of age required
}
\author{
Sara Hertog
}
