% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_helper_functions.R
\name{mid_period_pop}
\alias{mid_period_pop}
\title{Compute mid-period population of the population from population counts}
\usage{
mid_period_pop(pop_count_age_sex, period_length = 1)
}
\arguments{
\item{pop_count_age_sex}{data frame. columns "time_start, "sex", "age_start" and "value" where value is population counts}
}
\value{
a data frame with time_start, sex, age_start and mid-period population counts in the "value" field
}
\description{
This function computes mid period population by sex and age from population counts by sex and age
}
\details{
simple average of population by age and sex at two successive points in time
}
\author{
Sara Hertog
}
