% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_net_count_df.R
\name{mig_net_count_age_sex}
\alias{mig_net_count_age_sex}
\alias{mig_net_count_age_sex.data.frame}
\alias{mig_net_count_age_sex.mig_net_prop_age_sex}
\alias{mig_net_count_age_sex.ccmpp_input_list}
\alias{mig_net_count_age_sex<-}
\title{Constructor for class \code{mig_net_count_age_sex}}
\usage{
mig_net_count_age_sex(x, ...)

\method{mig_net_count_age_sex}{data.frame}(x, value_scale = attr(x, "value_scale"), ...)

\method{mig_net_count_age_sex}{mig_net_prop_age_sex}(
  x,
  pop_count_age_sex,
  value_scale_pop_count = attr(pop_count_age_sex, "value_scale"),
  ...
)

\method{mig_net_count_age_sex}{ccmpp_input_list}(x)

mig_net_count_age_sex(x) <- value
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Details}).}

\item{...}{Passed to the low-level constructor.}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}

\item{pop_count_age_sex}{For the \code{mig_net_prop_age_sex}
method, an object that can be coreced to a
\code{\link{ccmpp_input_df}} object (e.g., a
\code{data.frame}), holding population counts from which to
calculate the migration counts. The \code{value_type} must be
\dQuote{count} and the \code{value_scale}s must match.#'}

\item{value_scale_pop_count}{The scale of the \code{value} column
in \code{pop_count_age_sex}. If unspecified, defaults to the
\dQuote{value_scale} attribute of \code{pop_count_age_sex} if
non-\code{NULL}, otherwise \code{1}.}
}
\value{
An object of class \code{mig_net_count_age_sex}.
}
\description{
\code{mig_net_count_age_sex} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{Within year and sex, age must start at 0.}}
}
\details{
Methods are defined for \code{\link{data.frame}}s and
\code{\link{ccmpp_input_list}}s, and possibly other objects as
well. The \code{data.frame} method \dQuote{constructs} an object
from \code{x}. The \code{ccmpp_input_list} method \dQuote{extracts}
an object from \code{x}. There is also a replacement function which
complements the extraction methods.
}
\section{Transformation of migration proportions}{

The \code{mig_net_prop_age_sex} method has the mandatory argument
\code{pop_count_age_sex}. It is assumed that these already include
migration, and that this migration is that given by the proportions
passed in as argument \code{x}. The function then simply multiplies
the proportions (\code{x}) by the population counts
(\code{pop_count_age_sex}). Contrast this with the
\code{\link{mig_net_prop_age_sex}} replacement method for
\code{ccmpp_input_list}s which replaces the migration \emph{counts}
in the CCMPP input list such that they yield the specified
proportions.
}

\seealso{
\code{\link{validate_ccmppWPP_object}} for object
    validation, \code{\link{ccmpp_input_df}} for the class from
    which this one inherits. See \code{\link{mig_net_prop_age_sex}}
    for conversion between migration counts and proportions.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_count_age_sex}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mort_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_pop_count_age_sex_reference}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{death_probability_age_sex}()},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{mort_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
