% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_net_count_input_df.R
\name{mig_net_count_input_df}
\alias{mig_net_count_input_df}
\title{Constructor for class \code{mig_net_count_input_df}}
\usage{
mig_net_count_input_df(
  x,
  dimensions = attr(x, "dimensions"),
  age_span = attr(x, "age_span"),
  time_span = attr(x, "time_span"),
  ...
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{age_span}{Scalar indicating the span of the age groups. If
\code{NULL} and a column \dQuote{age_span} is present in
\code{x} an attempt will be made to infer it from that column.}

\item{time_span}{Scalar indicating the span of the time periods;
similar behaviour to \code{age_span}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{mig_net_count_input_df}.
}
\description{
\code{mig_net_count_input_df} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{Within year and sex, age must start at 0.}}
}
\seealso{
\code{\link{validate_ccmpp_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{fert_rate_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{srb_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
