% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_net_prop_df.R
\name{mig_net_prop_age_sex}
\alias{mig_net_prop_age_sex}
\alias{mig_net_prop_age_sex.data.frame}
\alias{mig_net_prop_age_sex.ccmpp_input_list}
\alias{mig_net_prop_age_sex.mig_net_count_age_sex}
\title{Constructor for class \code{mig_net_prop_age_sex}}
\usage{
mig_net_prop_age_sex(x, ...)

\method{mig_net_prop_age_sex}{data.frame}(x, ...)

\method{mig_net_prop_age_sex}{ccmpp_input_list}(x, ...)

\method{mig_net_prop_age_sex}{mig_net_count_age_sex}(x, pop_count_age_sex, ...)
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Usage}).}

\item{...}{Passed to the low-level constructor.}

\item{pop_count_age_sex}{For the \code{mig_net_count_age_sex}
method, an object that can be coreced to a
\code{\link{ccmpp_input_df}} object (e.g., a
\code{data.frame}), holding population counts from which to
calculate the migration proportions. The \code{value_type} must
be \dQuote{count} and the \code{value_scale}s must match.}
}
\value{
An object of class \code{mig_net_prop_age_sex} with the
    same \code{\link{value_scale}} as \code{x}.
}
\description{
\code{mig_net_prop_age_sex} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{prop}.}
  \item{Within year and sex, age must start at 0.}}
\code{mig_net_prop_age_sex} objects are not a required element of
\code{\link{ccmpp_input_list}}.
}
\details{
This function is generic with methods for \code{data.frame},
\code{ccmpp_input_list} and \code{mig_net_count_age_sex}. The
latter two methods will compute the proportions from counts of
migration and a baseline population by simple division of migration
counts by population counts, matched by age, sex, and
time. Population counts are supplied as a data frame via argument
\code{pop_count_age_sex}. The method for \code{ccmpp_input_list}s
uses \code{\link{project_ccmpp_loop_over_time}} to generate the
population counts from the inputs supplied via argument \code{x}.
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_count_age_sex}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mort_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{death_probability_age_sex}()},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{mort_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
