% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_parameter_df.R
\name{mig_parameter}
\alias{mig_parameter}
\alias{mig_parameter.data.frame}
\alias{mig_parameter.ccmpp_input_list}
\alias{mig_parameter<-}
\title{Constructor for class \code{mig_parameter}}
\usage{
mig_parameter(x, ...)

\method{mig_parameter}{data.frame}(x)

\method{mig_parameter}{ccmpp_input_list}(x)

mig_parameter(x) <- value
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Details}).}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{mig_parameter}.
}
\description{
\code{mig_parameter} is a subclass of
\code{\link{ccmpp_input_df}}. It contains
}
\details{
Methods are defined for \code{\link{data.frame}}s and
\code{\link{ccmpp_input_list}}s, and possibly other objects as
well. The \code{data.frame} method \dQuote{constructs} an object
from \code{x}. The \code{ccmpp_input_list} method \dQuote{extracts}
an object from \code{x}. There is also a replacement function which
complements the extraction methods.
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_count_age_sex}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mort_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_pop_count_age_sex_reference}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{death_probability_age_sex}()},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mort_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
