% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{new_ccmpp_input_list}
\alias{new_ccmpp_input_list}
\title{Low-level constructor for class \code{ccmpp_input_list}.}
\usage{
new_ccmpp_input_list(
  x,
  age_span = double(),
  time_span = double(),
  non_zero_fert_ages = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{ccmpp_input_list}.
}
\description{
Creates an object of class \code{ccmpp_input_list}. Minimal
checks are done; for interactive use see
\code{\link{ccmpp_input_list}}.

This function is not exported. The user-level constructor is
\code{\link{ccmpp_input_list}}.
}
\seealso{
ccmpp_input_list
}
\author{
Mark Wheldon
}
