% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{new_demog_change_component_df}
\alias{new_demog_change_component_df}
\title{Low-level constructor for class \code{demog_change_component_df}.}
\usage{
new_demog_change_component_df(
  x,
  dimensions = character(),
  value_type = character(),
  value_scale = double(),
  ...,
  class = character()
)
}
\arguments{
\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{demog_change_component_df}.
}
\description{
Creates an object of class \code{demog_change_component_df}. Minimal
checks are done; for interactive use see
\code{\link{demog_change_component_df}}.
}
\details{
This function is not exported. The user-level constructor is
\code{\link{demog_change_component_df}}.
}
\section{Default values for attributes}{

\subsection{Value type}{
If \code{value_type} is \code{NULL} the following defaults are employed:
\itemize{
\item If all elements of the \code{value} column are \code{NA}, \code{value_type} is set to \code{NA}.
\item If all elements of the \code{value} column are \code{numeric}, \code{value_type} is set to \code{real}.
\item If all elements of the \code{value} column are \code{character}, \code{value_type} is set to \code{categorical}.
}
}
\subsection{Value scale}{
If \code{value_scale} is \code{NULL} the following defaults are employed:
\itemize{
\item If \code{value_type} is such that \code{NA}s are sensible, \code{value_scale} is set to \code{NA}.
\item If \code{NA}s are not sensible given \code{value_type}, \code{value_scale} is set to \code{1}.
}
}
}

\seealso{
demog_change_component_df
}
\author{
Mark Wheldon
}
