% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{new_fert_rate_age_f}
\alias{new_fert_rate_age_f}
\title{Low-level constructor for class \code{fert_rate_age_f}.}
\usage{
new_fert_rate_age_f(
  x,
  age_span = double(),
  time_span = double(),
  non_zero_fert_ages = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{fert_rate_age_f}.
}
\description{
Creates an object of class \code{fert_rate_age_f}. Minimal
checks are done; for interactive use see
\code{\link{fert_rate_age_f}}.

This function is not exported. The user-level constructor is
\code{\link{fert_rate_age_f}}.
}
\seealso{
fert_rate_age_f
}
\author{
Mark Wheldon
}
