% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{new_fert_rate_input_df}
\alias{new_fert_rate_input_df}
\title{Low-level constructor for class \code{fert_rate_input_df}.}
\usage{
new_fert_rate_input_df(
  x,
  age_span = double(),
  time_span = double(),
  non_zero_fert_ages = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{age_span}{Scalar indicating the span of the age groups. If
\code{NULL} and a column \dQuote{age_span} is present in
\code{x} an attempt will be made to infer it from that column.}

\item{time_span}{Scalar indicating the span of the time periods;
similar behaviour to \code{age_span}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{fert_rate_input_df}.
}
\description{
Creates an object of class \code{fert_rate_input_df}. Minimal
checks are done; for interactive use see
\code{\link{fert_rate_input_df}}.

This function is not exported. The user-level constructor is
\code{\link{fert_rate_input_df}}.
}
\seealso{
fert_rate_input_df
}
\author{
Mark Wheldon
}
\concept{fert_rate_input_df class non-exported functions}
