% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_life_table_df.R
\name{new_life_table_age_sex}
\alias{new_life_table_age_sex}
\title{Low-level constructor for class \code{life_table_age_sex}.}
\usage{
new_life_table_age_sex(
  x,
  age_span = double(),
  time_span = double(),
  dimensions = get_req_dimensions_for_ccmpp_in_out_classes("life_table_age_sex"),
  value_type = get_value_types_for_ccmpp_in_out_classes("life_table_age_sex"),
  value_scale = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{life_table_age_sex}.
}
\description{
Creates an object of class \code{life_table_age_sex}. Minimal
checks are done; for interactive use see
\code{\link{life_table_age_sex}}.

This function is not exported. The user-level constructor is
\code{\link{life_table_age_sex}}.
}
\seealso{
life_table_age_sex
}
\author{
Mark Wheldon
}
