% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_life_table_input_df.R
\name{new_life_table_input_df}
\alias{new_life_table_input_df}
\title{Low-level constructor for class \code{life_table_input_df}.}
\usage{
new_life_table_input_df(
  x,
  age_span = double(),
  time_span = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{life_table_input_df}.
}
\description{
Creates an object of class \code{life_table_input_df}. Minimal
checks are done; for interactive use see
\code{\link{life_table_input_df}}.

This function is not exported. The user-level constructor is
\code{\link{life_table_input_df}}.
}
\seealso{
life_table_input_df
}
\author{
Mark Wheldon
}
