% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_net_count_tot_input_df.R
\name{new_mig_net_count_tot_b}
\alias{new_mig_net_count_tot_b}
\title{Low-level constructor for class \code{mig_net_count_tot_b}.}
\usage{
new_mig_net_count_tot_b(x, time_span = double(), ..., class = character())
}
\arguments{
\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{mig_net_count_tot_b}.
}
\description{
Creates an object of class \code{mig_net_count_tot_b}. Minimal
checks are done; for interactive use see
\code{\link{mig_net_count_tot_b}}.

This function is not exported. The user-level constructor is
\code{\link{mig_net_count_tot_b}}.
}
\seealso{
mig_net_count_tot_b
}
\author{
Mark Wheldon
}
