% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_net_rate_input_df.R
\name{new_mig_net_rate_age_sex}
\alias{new_mig_net_rate_age_sex}
\title{Low-level constructor for class \code{mig_net_rate_age_sex}.}
\usage{
new_mig_net_rate_age_sex(
  x,
  age_span = double(),
  time_span = double(),
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{mig_net_rate_age_sex}.
}
\description{
Creates an object of class \code{mig_net_rate_age_sex}. Minimal
checks are done; for interactive use see
\code{\link{mig_net_rate_age_sex}}.

This function is not exported. The user-level constructor is
\code{\link{mig_net_rate_age_sex}}.
}
\seealso{
mig_net_rate_age_sex
}
\author{
Mark Wheldon
}
