% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_parameter_input_df.R
\name{new_mig_parameter}
\alias{new_mig_parameter}
\title{Low-level constructor for class \code{mig_parameter}.}
\usage{
new_mig_parameter(x, time_span = double(), ..., class = character())
}
\arguments{
\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{mig_parameter}.
}
\description{
Creates an object of class \code{mig_parameter}. Minimal
checks are done; for interactive use see
\code{\link{mig_parameter}}.

This function is not exported. The user-level constructor is
\code{\link{mig_parameter}}.
}
\seealso{
mig_parameter
}
\author{
Mark Wheldon
}
