% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_srb_input_df.R
\name{new_srb}
\alias{new_srb}
\title{Low-level constructor for class \code{srb}.}
\usage{
new_srb(x, time_span = double(), ..., class = character())
}
\arguments{
\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{srb}.
}
\description{
Creates an object of class \code{srb}. Minimal
checks are done; for interactive use see
\code{\link{srb}}.

This function is not exported. The user-level constructor is
\code{\link{srb}}.
}
\seealso{
srb
}
\author{
Mark Wheldon
}
