% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_srb_input_df.R
\name{new_srb_input_df}
\alias{new_srb_input_df}
\title{Low-level constructor for class \code{srb_input_df}.}
\usage{
new_srb_input_df(x, time_span = double(), ..., class = character())
}
\arguments{
\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{srb_input_df}.
}
\description{
Creates an object of class \code{srb_input_df}. Minimal
checks are done; for interactive use see
\code{\link{srb_input_df}}.

This function is not exported. The user-level constructor is
\code{\link{srb_input_df}}.
}
\seealso{
srb_input_df
}
\author{
Mark Wheldon
}
