% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_survival_ratio_input_df.R
\name{new_survival_ratio_input_df}
\alias{new_survival_ratio_input_df}
\title{Low-level constructor for class \code{survival_ratio_input_df}.}
\usage{
new_survival_ratio_input_df(
  x,
  age_span = double(),
  time_span = double(),
  non_zero_fert_ages = double(),
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{survival_ratio_input_df}.
}
\description{
Creates an object of class \code{survival_ratio_input_df}. Minimal
checks are done; for interactive use see
\code{\link{survival_ratio_input_df}}.

This function is not exported. The user-level constructor is
\code{\link{survival_ratio_input_df}}.
}
\seealso{
survival_ratio_input_df
}
\author{
Mark Wheldon
}
