% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_vital_rate_df.R
\name{new_vital_rate_df}
\alias{new_vital_rate_df}
\title{Low-level constructor for class \code{vital_rate_df}.}
\usage{
new_vital_rate_df(x, age_span, time_span, dimensions, ..., class = character())
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}
}
\value{
An object of class \code{vital_rate_df}.
}
\description{
Creates an object of class \code{vital_rate_df}. Minimal
checks are done; for interactive use see
\code{\link{validate_vital_rate_df}}.

This function is not exported. The user-level constructor is
\code{\link{vital_rate_df}}.
}
\seealso{
vital_rate_df
}
\author{
Mark Wheldon
}
\concept{vital_rate_df class non-exported functions}
