% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{new_vital_rate_df_time_age_sex}
\alias{new_vital_rate_df_time_age_sex}
\title{Low-level constructor for class \code{vital_rate_df_time_age_sex}.}
\usage{
new_vital_rate_df_time_age_sex(
  x,
  age_span,
  time_span,
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{vital_rate_df_time_age_sex}.
}
\description{
Creates an object of class \code{vital_rate_df_time_age_sex}. Minimal
checks are done; for interactive use see
\code{\link{validate_vital_rate_df_time_age_sex}}.

This function is not exported. The user-level constructor is
\code{\link{vital_rate_df_time_age_sex}}.
}
\seealso{
vital_rate_df_time_age_sex
}
\author{
Mark Wheldon
}
\concept{vital_rate_df_time_age_sex class non-exported functions}
