% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{parse_census_years_ranges}
\alias{parse_census_years_ranges}
\title{Parse vector of years containing ranges}
\usage{
parse_census_years_ranges(times)
}
\arguments{
\item{times}{A vector of times (probably years), numeric or
character, possibly with ranges in the format
\code{"xxxx-yyyy"} (see \dQuote{Description}).}
}
\value{
A numeric vector of times.
}
\description{
Parses a character vector of times (usually years). Single times
are left as-is and converted to numeric (via
\code{\link{as.numeric}}). Ranges in the form \code{"xxxx-yyyy"}
(e.g., as returned by \code{link{get_census_years}}) are replaced
by their midpoints, rounded down to the nearest integer, and then
converted. This function is called by
\code{\link{DDextract_to_pop_count_age_sex_reference}}. See that
function's help file for additional notes.
}
\seealso{
Other DDSQL_census_years: 
\code{\link{get_census_years}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_census_years}
