% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData_extract_post_process.R
\name{parse_pop_count_reference_times_ranges}
\alias{parse_pop_count_reference_times_ranges}
\title{Parse vector of years containing ranges}
\usage{
parse_pop_count_reference_times_ranges(times)
}
\arguments{
\item{times}{A vector of times (probably years), numeric or
character, possibly with ranges in the format
\code{"xxxx-yyyy"} (see \dQuote{Description}).}
}
\value{
A numeric vector of times.
}
\description{
Parses a character vector of times (usually years). Single times
are left as-is and converted to numeric (via
\code{\link{as.numeric}}). Ranges in the form \code{"xxxx-yyyy"}
(e.g., as returned by \code{link{get_census_years}}) are replaced
by their midpoints and then converted. This function is primarily
intended to be called via
\code{\link{get_pop_count_age_sex_reference}}. See that function's
help file for more details.
}
\seealso{
Other DDSQL_extract: 
\code{\link{get_census_years}()},
\code{\link{get_pop_count_age_sex_reference}()},
\code{\link{get_pop_count_reference_times}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
