% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_variants.R
\name{pasfr_given_tfr}
\alias{pasfr_given_tfr}
\title{Infer proportion age-specific fertility rates from tfr using a global normative model}
\usage{
pasfr_given_tfr(pasfr_initial, tfr_initial, tfr_given)
}
\arguments{
\item{pasfr_initial}{data frame. with age_start and age_span indicating age groups and initial pasfr in the value column}

\item{tfr_initial}{numeric. the initial level of total fertility}

\item{tfr_given}{numeric.  the projected level of total fertility for which we want to infer a pasfr}
}
\value{
a data frame with the inferred pasfr in the "value" field.
}
\description{
This function returns the proportional age-specific fertility rates given a projected level of
total fertility by converging the initial pasfr to a global model
}
\details{
accesses the global model in the data frame "pasfr_global_model"
}
\author{
Sara Hertog
}
