% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_variants.R
\name{pasfr_global_model}
\alias{pasfr_global_model}
\title{Compute global normative model of proportion age-specific fertility rates}
\usage{
pasfr_global_model(
  tfr_all_locs,
  pasfr_all_locs,
  PASFRpattern,
  present_year = 2021,
  mac_criterion = 30
)
}
\arguments{
\item{tfr_all_locs}{data.frame. file path to RData file with annual TFR; long format with LocID, time_start and value}

\item{pasfr_all_locs}{data.frame. file path to RData file with 1x1 PASFR; long format with LocID, time_start and value, age_start from 10 to 54}

\item{present_year}{numeric. Last year of observed data?}

\item{mac_criterion}{numeric. Lower bound threshold for mean age at childbearing to determine includision in computing global norm}
}
\value{
a matrix of annual PASFR norms; age from 10 to 54 in row, year in columns
}
\description{
This function calls bayesPop functions to return the proportional age-specific fertility rates associated with a global norm
based on historical patterns of TFR and PASFR
}
\details{
calls embedded functions from bayesPop
}
\author{
Sara Hertog
}
