% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{plot.demog_change_component_df}
\alias{plot.demog_change_component_df}
\title{Plot \code{demog_change_component_df} objects}
\usage{
\method{plot}{demog_change_component_df}(x, type = c("line", "point", "both"), ...)
}
\arguments{
\item{x}{An object inheriting from class \code{demog_change_component_df}.}

\item{type}{The type of plot to produce, e.g., \dQuote{line}.}

\item{...}{Passed on to \code{\link[ggplot2]{ggplot}}.}
}
\value{
A \code{ggplot2} plot object.The function is mainly called for
    the side effect of producing a plot.
}
\description{
This is the S3 method for the generic \code{\link{plot}}
function. It requires \pkg{ggplot2}.
}
\author{
Mark Wheldon
}
