% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_output_pop_count_age_sex_df.R
\name{pop_count_age_sex}
\alias{pop_count_age_sex}
\title{Constructor for class \code{pop_count_age_sex}}
\usage{
pop_count_age_sex(x, ...)
}
\arguments{
\item{x}{A \code{\link{base::data.frame}} or
\code{link{ccmpp_input_list}} object.}

\item{...}{Passed to the low-level constructor.}

\item{keep_baseline}{Logical, \code{link{ccmpp_input_list}} method:
should the baseline population counts be included in the
output?}
}
\value{
An object of class \code{pop_count_age_sex}.
}
\description{
\code{pop_count_age_sex} is a subclass of
\code{\link{ccmpp_output_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{count}.}
  \item{Within year and sex, age must start at 0.}}
It is intended as a class that can be used for the population
counts produced by the CCMPP. There are methods for
\code{\link{base::data.frame}}s and \code{link{ccmpp_input_list}}s.
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{ccmpp_output_df}} for the class from which this one
    inherits.

Other ccmpp_output_objects: 
\code{\link{ccmpp_output_df}()},
\code{\link{coerce_ccmpp_output_df}},
\code{\link{coerce_pop_count_age_sex}}
}
\author{
Mark Wheldon
}
\concept{ccmpp_output_objects}
