% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_pop_count_reference.R
\name{pop_count_age_sex_reference}
\alias{pop_count_age_sex_reference}
\title{Constructor for class \code{pop_count_age_sex_reference}}
\usage{
pop_count_age_sex_reference(x, value_scale = attr(x, "value_scale"))
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Details}).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{pop_count_age_sex_reference}.
}
\description{
\code{pop_count_age_sex_reference} is a subclass of
\code{\link{demog_change_component_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{time_span} attribute equals 0.}
  \item{\code{value_type} attribute equals \dQuote{count}.}}
}
\details{
Methods are defined for \code{\link{data.frame}}s and possibly
other objects as well. The \code{data.frame} method
\dQuote{constructs} an object from \code{x}.
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{demog_change_component_df}} for the class from which this one
    inherits.
}
\author{
Mark Wheldon
}
