% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_ccmpp_input_df.R
\name{print.ccmpp_input_df}
\alias{print.ccmpp_input_df}
\title{Print Values of a \code{ccmpp_input_df}}
\usage{
\method{print}{ccmpp_input_df}(
  x,
  ...,
  n = min(6L, nrow(x)),
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = FALSE,
  na.print = ".",
  print_what = c("info", "table")
)
}
\arguments{
\item{x}{object of class \code{data.frame}.}

\item{...}{optional arguments to \code{print} or \code{plot} methods.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{na.print}{character string (or \code{NULL}) indicating
    \code{\link[base]{NA}} values in printed output, see
    \code{\link[base]{print.default}}.}
}
\description{
This is a method for the generic \code{\link{base::print}} function. Only
the first \code{n} rows are printed for convenience (by default). If all rows are desired use
\code{as.data.frame(x)} or see the definition of argument \code{n}.
}
\details{
Entries in columns \code{age_span} and \code{time_span} are
rendered as \dQuote{\verb{.}} to emphasize that these values are
equal over rows and columns. \emph{Note:} This is merely for
printing; the columns are still present with their actual numerical
values, i.e., \code{unique(x$time_span) == time_span(x)} is
\code{TRUE} and similarly for age span.
}
\seealso{
print.demog_change_component_df
}
\author{
Mark Wheldon
}
