% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_ccmpp_input_list.R
\name{print.ccmpp_input_list}
\alias{print.ccmpp_input_list}
\title{Print Values of a \code{ccmpp_input_list}}
\usage{
\method{print}{ccmpp_input_list}(
  x,
  ...,
  vsep,
  n = 6L,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = FALSE,
  print_what = c("info", "table")
)
}
\arguments{
\item{x}{An object of class \code{ccmpp_input_list}.}

\item{...}{optional arguments to \code{print} or \code{plot} methods.}

\item{n}{Integer controlling how many rows of \code{x} are
printed. Passed to \code{link{head}}; see the documentation of
that function for valid values and associated behaviours.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{print_what}{What should be printed?}
}
\description{
This is a method for the generic \code{\link{base::print}}
function. Only the first \code{n} rows of each component of
\code{x} are printed for convenience (by default).
}
\author{
Mark Wheldon
}
