% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_fert_rate_input_df.R
\name{print.fert_rate_age_f}
\alias{print.fert_rate_age_f}
\title{Print Values of a \code{fert_rate_age_f}}
\usage{
\method{print}{fert_rate_age_f}(
  x,
  ...,
  n = min(6L, nrow(x)),
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = FALSE,
  na.print = ".",
  print_what = c("info", "table")
)
}
\arguments{
\item{x}{object of class \code{data.frame}.}

\item{...}{optional arguments to \code{print} or \code{plot} methods.}

\item{n}{Integer controlling how many rows of \code{x} are
printed. Passed to \code{link{head}}; see the documentation of
that function for valid values and associated behaviours.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{na.print}{character string (or \code{NULL}) indicating
    \code{\link[base]{NA}} values in printed output, see
    \code{\link[base]{print.default}}.}

\item{print_what}{What should be printed?}
}
\description{
This is a method for the generic \code{\link{base::print}} function. Only
\code{n} rows are printed for convenience. If all rows are desired use
\code{as.data.frame(x)} or see the definition of argument \code{n}.
}
\details{
As well as inheriting the behaviour of
\code{\link{print.ccmpp_input_df}}, entries in the \code{value}
column corresponding to values of \code{age_start} \emph{not} in
\code{non_zero_fert_ages(x)} are rendered as
\dQuote{\verb{[zero]}}. At least some rows with non-zero fertility
rate ages will be shown if possible.
}
\seealso{
print.demog_change_component_df print.ccmpp_input_df fert_rate_age_f
}
\author{
Mark Wheldon
}
