% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{print.summary_demog_change_component_df}
\alias{print.summary_demog_change_component_df}
\title{Print a summary of a \code{demog_change_component_df}}
\usage{
\method{print}{summary_demog_change_component_df}(x, vsep, ...)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}

\item{...}{Currently not used.}
}
\value{
Called for its side-effect.
}
\description{

}
\seealso{
\code{\link{summary.demog_change_component_df}}.
}
\author{
Mark Wheldon
}
