% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.vital_rate_df_time_age_sex}
\alias{print.vital_rate_df_time_age_sex}
\title{Print Values of a \code{vital_rate_df_time_age_sex}}
\usage{
\method{print}{vital_rate_df_time_age_sex}(
  x,
  ...,
  n = 6L,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = TRUE,
  max = NULL
)
}
\arguments{
\item{x}{An object of class \code{vital_rate_df_time_age_sex}.}

\item{...}{optional arguments to \code{print} or \code{plot} methods.}

\item{n}{Integer controlling how many rows of \code{x} are
printed. Passed to \code{link{head}}; see the documentation of
that function for valid values and associated behaviours.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{max}{numeric or \code{NULL}, specifying the maximal number of
    entries to be printed.  By default, when \code{NULL},
    \code{\link[base]{getOption}("max.print")} used.}
}
\description{
This is a method for the generic \code{\link{print}} function. Only
the first \code{n} rows are printed for convenience (by default). If all rows are desired use
\code{as.data.frame(x)} or see the definition of argument \code{n}.
}
\author{
Mark Wheldon
}
