% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proj_pop_cpp}
\alias{proj_pop_cpp}
\title{Cohort component population projection (C++ implementation)}
\usage{
proj_pop_cpp(
  step_size,
  pop_count_age_m_t0,
  pop_count_age_f_t0,
  surv_prop_age_m,
  surv_prop_age_f,
  fert_rate_age,
  net_mig_count_age_m,
  net_mig_count_age_f,
  srb_tot = 1.05,
  mig_assumption = "even"
)
}
\arguments{
\item{step_size}{The size of the project step and age intervals. Usually this is 1 or 5.}

\item{pop_count_age_m_t0}{Vector of population counts by age, of males in the baseline year.}

\item{pop_count_age_f_t0}{Vector of population counts by age, of females in the baseline year.}

\item{surv_prop_age_m}{Vector of life table survival proportions by age, of males over the projection step.}

\item{surv_prop_age_f}{Vector of life table survival proportions by age, of females over the projection step.}

\item{fert_rate_age}{Vector of fertility rates by age of mother during the projection step.}

\item{net_mig_count_age_m}{Vector of net number of male migrants arriving, by age, during the projection step.}

\item{net_mig_count_age_f}{Vector of net number of female migrants arriving, by age, during the projection step.}

\item{srb_tot}{Average sex ratio at birth across all births during the projection step.}

\item{mig_assumption}{Treatment of net migration. See \dQuote{Migration Assumption}.}
}
\description{
The function projects the baseline population forward in time (one
step of length \code{step_size} years) by adding births and net
migrants and subtracting deaths. This implementation is written in
C++ (using the \pkg{Rcpp}
package).
}
\details{
The cohort component method of population projection (CCMPP) is a
deterministic method for projecting age-stratified population
counts forward in time, accounting for births, deaths, and net
migration. It is described in \cite{Preston, et al., (2001)},
Chapter 6, and many other places.
}
\section{Migration Assumption}{

The CCMPP is a discrete time approximation of a continuous time
process in which certain assumptions are made. Two such
assumptions about net migration are implemented:
\describe{
  \item{\dQuote{even}}{Net migrants are distributed across the projection interval by adding half of the net migration to the baseline population at the beginning of the projection interval therby exposing them to fertility and mortality. The remaining half is added at the end and are not exposed to mortality and fertility.}
  \item{\dQuote{end}}{All net migration is added at the end of the projection step, therefore none are exposed to mortality and fertility.}}
The default is \dQuote{even}; \dQuote{end} is included for compatibilty with some earlier implementations of the CCMPP.
}

\references{
Preston, S. H., Heuveline, P., and Guillot, M. (2001), \emph{Demography: Measuring and Modeling Population Processes}, Malden, Massachusetts: Blackwell.
}
\seealso{
\code{\link{project_z_by_z}}.
}
\author{
Mark Wheldon, based on \R code by Sara Hertog
}
