% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_abacus_z_by_z.R
\name{project_abacus_z_by_z}
\alias{project_abacus_z_by_z}
\title{R implementation of the one-step cohort component population projection approach from legacy Abacus}
\usage{
project_abacus_z_by_z(
  z = 5,
  pop_count_age_m_begin,
  pop_count_age_f_begin,
  survival_ratio_age_m,
  survival_ratio_age_f,
  fert_rate_age_f,
  srb,
  mig_net_count_age_m,
  mig_net_count_age_f,
  mig_assumption = c("end", "even")
)
}
\arguments{
\item{z}{width of age groups = length of one-step projection horizon (Abacus was 5 years only)}

\item{pop_count_age_m_begin}{a vector of male population counts at the base year (time 0) by age}

\item{pop_count_age_f_begin}{a vector of female population counts at the base year (time 0) by age}

\item{survival_ratio_age_m}{a vector of survival ratios by age for males from a life table referencing the period 0 to 0+z}

\item{survival_ratio_age_f}{a vector of survival ratios by age for females from a life table referencing the period 0 to 0+z}

\item{fert_rate_age_f}{a vector of age-specific fertility rates referencing the period 0 to 0+z}

\item{srb}{a scalar for the sex ratio at birth for the period 0 to 0+z (male births/female births)}

\item{mig_net_count_age_m}{a vector of net migration counts for males by age referencing the period 0 to 0+z}

\item{mig_net_count_age_f}{a vector of net migration counts for females by age referencing the period 0 to 0+z}

\item{mig_assumption}{a control for the migration assumption. This is what differentiates the Abacus implementation here from the
implementation in the project_ccmpp_z_by_z function.  In Abacus, "end" means migration is accounted at end of projection period, but 
there is still some exposure of female migrants to fertility and thus births are affected by the level of net migration.
"even" in Abacus accounts for migration evenly distributed over the period, but 
excludes beginning of period migrants from exposure to fertility. This is because Abacus first projects population and applies
the migration assumption and then computes births by exposing (beginning period females + end of period females)/2 to the age specific
fertility rates of the period.  We have deemed this to be not the ideal implementation of the migration assumption and have thus
adopted the cleaner approach in project_ccmpp_z_by_z for WPP 2021 
#'}
}
\value{
a list of objects including projected population by age and sex, deaths by age and sex,
births by mother's age, total births, and births by sex
}
\description{
This function takes a set of base population and demographic inputs for the period base year to base year+z years with age groups of width z and projects population forward z years according to the logic implemented in Abacus for WPP revisions 2019 and earlier.
}
\details{
This function will accept any width of age group and projection horizon as long as those are equal
e.g., use z=1 for a 1x1 projection of population by single year of age and 1 year projection horizon
or use z=5 for a 5x5 projection of population by 5-year age groups and 5-year projection horizon (this is what Abacus does)
}
\author{
Sara Hertog
}
