% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_ccmpp}
\alias{project_ccmpp}
\title{Cohort component population projection (C++ implementation)}
\usage{
project_ccmpp(
  step_size,
  P0M,
  P0F,
  SxM,
  SxF,
  asfr,
  NMxM,
  NMxF,
  srb = 1.05,
  migration_assumption = 1L
)
}
\description{
The cohort component method of population projection (CCMPP) is a
deterministic method for projecting age-stratified population
counts forward in time, accounting for births, deaths, and net
migration. It is described in \cite{Preston, et al., (2001)},
Chapter 6, and many other places.
}
\details{
This implementation is written in C++ and projects the baseline
population forward once \code{step_size} years.
}
\references{
Preston, S. H., Heuveline, P., and Guillot, M. (2001), \emph{Demography: Measuring and Modeling Population Processes}, Malden, Massachusetts: Blackwell.
}
\author{
Mark Wheldon, Sara Hertog
}
