% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ccmpp_loop_over_periods.R
\name{project_ccmpp_loop_over_periods}
\alias{project_ccmpp_loop_over_periods}
\title{Loop over periods for cohort component population projection}
\usage{
project_ccmpp_loop_over_periods(
  z = 1,
  year_start,
  year_stop,
  pop_base_input_m,
  pop_base_input_f,
  fert_input,
  srb_input,
  mort_input_m,
  mort_input_f,
  migration_type,
  mig_input_m,
  mig_input_f,
  migration_assumption
)
}
\arguments{
\item{z}{numeric. width of age groups = length of one-step projection horizon}

\item{year_start}{numeric. base year of projection}

\item{year_stop}{numeric. end year for projection}

\item{pop_base_input_m}{numeric. vector of male population counts at year=year_start}

\item{pop_base_input_f}{numeric. vector of female population counts at year=year_start}

\item{fert_input}{data frame. with columns "year", "age" and "value" with value containing age-specific 
fertility rates}

\item{mort_input_m}{data frame. with columns "year", "age" and "value" with value containing time and 
age-specific life table survival ratios for males}

\item{mort_input_f}{data frame. with columns "year", "age" and "value" with value containing time and 
age-specific life table survival ratios for females}

\item{migration_type}{character. "counts" if migration inputs are counts of net migrants; "rates" if migration 
inputs are as a proportion of population by age and sex.}

\item{mig_input_m}{data frame. with columns "year", "age" and "value" with value containing time and 
age-specific net migration values for males. Can be expressed as net counts or as a proportion of population (rates).}

\item{mig_input_f}{data frame. with columns "year", "age" and "value" with value containing time and 
age-specific net migration values for femles. Can be expressed as net counts or as a proportion of population (rates).}

\item{migration_assumption}{data frame. with columns "year" and "value" where value is a character string 
indicating whether the project_ccmpp_z_by_z function should implement net migration "even"ly over period or at
"end" of period. 
#'}

\item{srb}{data frame. with columns "year" and "value" with value containing the sex-ratio at birth (M/F)}
}
\value{
a list of objects with step by step cohort-component population projection results, including population
by age and sex, period deaths by cohort and sex, births by age of the mother, total period births by sex, and net 
migrant counts by age and sex
}
\description{
This function takes a set of base population and time series of demographic inputs 
and carries out the cohort component population projection over the specified periods
}
\details{
This function takes a full set of base population and demographic inputs for a country over time and 
implements a cohort-component population projection with steps of length z by looping through the 
project_ccmpp_z_by_z() function.
}
\author{
Sara Hertog
}
