% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ccmpp_loop_over_time.R
\name{project_ccmpp_loop_over_time}
\alias{project_ccmpp_loop_over_time}
\title{Loop over periods for cohort component population projection}
\usage{
project_ccmpp_loop_over_time(indata)
}
\arguments{
\item{indata}{list. ccmpp_input list of data frames}
}
\value{
a list of objects with step by step cohort-component population projection results, including population
by age and sex, period deaths by cohort and sex, births by age of the mother, total period births by sex, and net
migrant counts by age and sex
}
\description{
This function takes a set of base population and time series of demographic inputs
and carries out the cohort component population projection over the specified time periods
}
\details{
This function takes a full set of base population and demographic inputs for a country over time and
implements a cohort-component population projection with steps of length z by looping through the
project_ccmpp_z_by_z() function.
}
\author{
Sara Hertog
}
