% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ccmpp_z_by_z_backwards.R
\name{project_ccmpp_z_by_z_backwards}
\alias{project_ccmpp_z_by_z_backwards}
\title{One-step backwards cohort component population projection with flexible migration assumption}
\usage{
project_ccmpp_z_by_z_backwards(
  z = 1,
  pop_count_age_m_start,
  pop_count_age_f_start,
  survival_ratio_age_m,
  survival_ratio_age_f,
  nLx_age_m,
  nLx_age_f,
  mig_net_count_age_m,
  mig_net_count_age_f,
  mig_assumption = c("end", "even")
)
}
\arguments{
\item{z}{numeric. width of age groups = length of one-step projection horizon}

\item{pop_count_age_m_start}{numeric. a vector of male population counts at the start of the projection step by age}

\item{pop_count_age_f_start}{numeric. a vector of female population counts at the start of the projection step by age}

\item{survival_ratio_age_m}{numeric. a vector of survival ratios by age for males from a life table referencing the 
period time_start-z to time_start}

\item{survival_ratio_age_f}{numeric. a vector of survival ratios by age for females from a life table referencing the 
period time_start-z to time_start}

\item{nLx_age_m}{numeric. a vector of life table person-years for males from a life table referencing period 
time_start-z to time_start}

\item{nLx_age_f}{numeric. a vector of life table person-years for females from a life table referencing period 
time_start-z to time_start}

\item{mig_net_count_age_m}{numeric. a vector of net migration counts for males by age referencing the period 
time_start-z to time_start}

\item{mig_net_count_age_f}{numeric. a vector of net migration counts for females by age referencing the period 
time_start-z to time_start}

\item{mig_assumption}{charater. a control for the migration assumption. "end" means migration is accounted at end of projection 
period and has no influence on deaths during the period. "even" accounts for migration evenly distributed 
over the period and thus does influence deaths}
}
\value{
a list of objects including projected population by age and sex and deaths by cohort and sex
}
\description{
This function takes a set of starting population and demographic inputs for the period time_start-z 
to time_start with age groups of width z and projects population backwards z units of time.
}
\details{
This function will accept any width of age group and projection horizon as long as those are equal
e.g., use z=1 for a 1x1 projection of population by single year of age and 1 year projection horizon
or use z=5 for a 5x5 projection of population by 5-year age groups and 5-year projection horizon
}
\author{
Sara Hertog
}
