% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{rbind.demog_change_component_df}
\alias{rbind.demog_change_component_df}
\title{Combine \code{demog_change_component_df} objects}
\usage{
\method{rbind}{demog_change_component_df}(
  ...,
  deparse.level = 1,
  make.row.names = TRUE,
  stringsAsFactors = default.stringsAsFactors(),
  factor.exclude = TRUE
)
}
\arguments{
\item{...}{(generalized) vectors or matrices.  These can be given as named
    arguments.  Other \R objects may be coerced as appropriate, or S4
    methods may be used: see sections \sQuote{Details} and
    \sQuote{Value}.  (For the \code{"data.frame"} method of \code{cbind}
    these can be further arguments to \code{\link[base]{data.frame}} such as
    \code{stringsAsFactors}.)}

\item{deparse.level}{integer controlling the construction of labels in
    the case of non-matrix-like arguments (for the default method):\cr
    \code{deparse.level = 0} constructs no labels; the default,\cr
    \code{deparse.level = 1 or 2} constructs labels from the argument
    names, see the \sQuote{Value} section below.}

\item{make.row.names}{(only for data frame method:) logical
    indicating if unique and valid \code{\link[base]{row.names}} should be
    constructed from the arguments.}

\item{stringsAsFactors}{logical, passed to \code{\link[base]{as.data.frame}};
    only has an effect when the \code{\dots} arguments contain a
    (non-\code{data.frame}) \code{\link[base]{character}}.}

\item{factor.exclude}{if the data frames contain factors, the default
    \code{TRUE} ensures that \code{NA} levels of factors are kept, see
    \href{https://bugs.R-project.org/bugzilla3/show_bug.cgi?id=17562}{PR#17562} and the \sQuote{Data frame methods}.  In \R versions up
    to 3.6.x, \code{factor.exclude = NA} has been implicitly hardcoded
    (\R <= 3.6.0) or the default (\R = 3.6.x, x >= 1).}
}
\value{
A data frame.
}
\description{
This is an S3 method for \code{\link{base::rbind}} for objects that
inherit from \code{demog_change_component_df}. The special classes
are dropped. If called interactively a warning to this effect is
issued.
}
\author{
Mark Wheldon
}
