% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_validators_helpers.R
\name{sexes_unequal}
\alias{sexes_unequal}
\title{TRUE' quadruples it}
\usage{
sexes_unequal(
  x,
  x_name = "x",
  tolerance = 1e-06,
  scale = NULL,
  check_by_time = FALSE
)
}
\description{
TRUE' quadruples it
}
