% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_subclasses_access_attr.R
\name{span}
\alias{span}
\alias{span.ccmpp_input_df}
\alias{span.ccmpp_output_df}
\title{Return the \dQuote{span} of a \code{ccmpp_input_df} or \code{ccmpp_output_df}}
\usage{
span(x, ...)

\method{span}{ccmpp_input_df}(x, ...)

\method{span}{ccmpp_output_df}(x, ...)
}
\arguments{
\item{x}{An object inheriting from \code{ccmpp_input_df}.}

\item{...}{Passed to other methods.}
}
\value{
The span of \code{x}.
}
\description{
Returns the common \dQuote{age_span} and \dQuote{time_span} of
objects inheriting from \code{ccmpp_input_df} or
\code{ccmpp_output_df}. To be a valid member of this class, these
two spans must be equal. This function provides both a convenient
check for this equality and a more logical way of accessing the
common span. To change the span, see \code{\link{span<-}}.
}
\seealso{
\code{\link{extract_demog_change_component_attributes}},
    the assignment version for changing the span
    \code{\link{span<-}}
}
\author{
Mark Wheldon
}
