% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_srb_input_df.R
\name{srb_input_df}
\alias{srb_input_df}
\title{Constructor for class \code{srb_input_df}}
\usage{
srb_input_df(x, time_span = attr(x, "time_span"))
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}
}
\value{
An object of class \code{srb_input_df}.
}
\description{
\code{srb_input_df} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes three additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{ratio}.}
  \item{There can be no 'age' dimension.}
  \item{There can be no 'sex' dimension.}}
}
\seealso{
\code{\link{validate_ccmpp_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{fert_rate_input_df}()},
\code{\link{life_table_input_df}()},
\code{\link{mig_net_count_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
