% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{subset.demog_change_component_df}
\alias{subset.demog_change_component_df}
\title{Subsetting \code{demog_change_component_df}s}
\usage{
\method{subset}{demog_change_component_df}(x, ...)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}

\item{...}{Passed to \code{\link{subset.data.frame}}. See the help
file there for required arguments and defaults.}
}
\value{
A \code{demog_change_component_df} if valid after
    subsetting, otherwise a \code{data.frame}.
}
\description{
A method for \code{\link{base::subset}} for objects of class
\code{demog_change_component_df}s.
}
\details{
The arguments and defaults are the same as for the
\code{data.frame} method. Notably, the actual column names must be
used, (e.g., \code{time_start}, \emph{not} \code{time}). To subset
using conceptual dimension names see, e.g.,
\code{\link{subset_time}}.
}
\author{
Mark Wheldon
}
