% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_utility_functions.R
\name{subset_ccmpp_input_list}
\alias{subset_ccmpp_input_list}
\alias{subset_time.ccmpp_input_list}
\alias{subset_age.ccmpp_input_list}
\title{Subset by time, age, or sex}
\usage{
\method{subset_time}{ccmpp_input_list}(x, times, include = TRUE)

\method{subset_age}{ccmpp_input_list}(x, ages, include = TRUE)
}
\arguments{
\item{x}{An object to subset.}

\item{times, ages}{Vectors indicating the levels of
time, age, or sex to retain.}

\item{include}{Logical; should the rows corresponding to the values
supplied in the previous argument be those that are kept or
discarded?}
}
\value{
The object after subsetting if a valid member of the class;
    otherwise an error.
}
\description{
These functions subset the component data frames of objects
inheriting from \code{\link{ccmpp_input_list}} by \dQuote{time} or
\dQuote{age} and retain the class if valid (see
\dQuote{Details}). Subsetting such that only one level of the
subset dimension is retained will drop the dimension if \code{drop
= TRUE}, otherwise it is retained (default). Subsetting a
\code{ccmpp_input_list} by \dQuote{indicator} or
\dQuote{sex} cannot produce a valid member of the class.
}
\details{
If the object returned after subsetting is a valid member of the
original class it will be returned (valid according to
\code{\link{validate_ccmppWPP_object}}). If it is not valid an error
will be signalled and nothing is returned.
}
\section{Note}{

These functions are not particularly efficient. For repeated
subsetting within, e.g., a for loop, it is better to use standard
subsetting operations on the data frame component and re-cast the
result as a classed object at the end if desired.
}

\seealso{
\code{link{subset_time}}, etc., for the S3 generic
    function; \code{link{subset_time.demog_change_component_df}}
    for the method for objects inheriting from
    \code{demog_change_component_df}.

Other subset_by_demographic_dimension: 
\code{\link{subset_time.death_count_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{subset_by_demographic_dimension}
