% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{subset_replace}
\alias{subset_replace}
\alias{[.vital_rate_df_time_age_sex}
\alias{[<-.vital_rate_df_time_age_sex}
\alias{$<-.vital_rate_df_time_age_sex}
\alias{[[<-.vital_rate_df_time_age_sex}
\title{Extract or Replace Parts of a \code{vital_rate_df_time_age_sex}}
\usage{
\method{[}{vital_rate_df_time_age_sex}(x, i, j, drop)

\method{[}{vital_rate_df_time_age_sex}(x, i, j) <- value

\method{$}{vital_rate_df_time_age_sex}(x, name) <- value

\method{[[}{vital_rate_df_time_age_sex}(x, i, j) <- value
}
\arguments{
\item{x}{data frame.}

\item{i}{elements to extract or replace.  For \code{[} and
    \code{[[}, these are \code{numeric} or \code{character} or, for
    \code{[} only, empty.  Numeric values are coerced to integer as if
    by \code{\link[base]{as.integer}}.  For replacement by \code{[}, a logical
    matrix is allowed.}

\item{j}{elements to extract or replace.  For \code{[} and
    \code{[[}, these are \code{numeric} or \code{character} or, for
    \code{[} only, empty.  Numeric values are coerced to integer as if
    by \code{\link[base]{as.integer}}.  For replacement by \code{[}, a logical
    matrix is allowed.}

\item{drop}{logical.  If \code{TRUE} the result is coerced to the
    lowest possible dimension.  The default is to drop if only one
    column is left, but \bold{not} to drop if only one row is left.}

\item{value}{A suitable replacement value: it will be repeated a whole
    number of times if necessary and it may be coerced: see the
    Coercion section.  If \code{NULL}, deletes the column if a single
    column is selected.}

\item{name}{
    A literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).}
}
\value{
Extraction could return a \code{data.frame},
    one-dimensional vector, or scalar depending on the dimensions
    of the extracted values. Replacement will
    attempt to return a \code{vital_rate_df_time_age_sex}.
}
\description{
These are methods for the subset operator and its companion
replacement function for objects of class
\code{vital_rate_df_time_age_sex}. The return values differ and,
importantly, subsetting does not return an object of class
\code{vital_rate_df_time_age_sex} (see \dQuote{Details}).
}
\details{
The object resulting from a subset operation via \code{`[`} will no
longer inherit from class \code{vital_rate_df_time_age_sex} and the
attributes specific to that class will be lost. To create a
\code{vital_rate_time_age_sex_df} object from a subset call
\code{\link{vital_rate_df_time_age_sex}} explicitly (e.g.,
\code{vital_rate_df_time_age_sex(x[x$time_start == 1960,])}).

Replacement via \code{`[<-`} will attempt to return a
\code{vital_rate_matrix}. The user is responsible for making sure
the new object is a valid member of this class (see
\code{\link{vital_rate_df_time_age_sex}}). If the result is not a valid
member an error will be signalled.
}
\seealso{
\code{\link{vital_rate_df_time_age_sex}}.
}
\author{
Mark Wheldon
}
