% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_last_column.R
\name{sum_last_column}
\alias{sum_last_column}
\title{Compute sum of one column of a data frame by the values of the other columns in the data frame}
\usage{
sum_last_column(indata)
}
\arguments{
\item{indata}{data frame. last (rightmost) column must be numeric.}
}
\value{
a data frame with same names as indata for which the last column is the computed sum of values in the last
column of indata
}
\description{
This is a general function that computes the sum of the last (rightmost) column of a data frame
indexed by the combinations of values of all of the preding columns of that data frame.
}
\author{
Sara Hertog
}
