% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{summary.demog_change_component_df}
\alias{summary.demog_change_component_df}
\title{Summarize a \code{demog_change_component_df}}
\usage{
\method{summary}{demog_change_component_df}(
  object,
  maxsum = 7,
  digits = max(3, getOption("digits") - 3),
  vsep,
  ...
)
}
\arguments{
\item{object}{An object of class \code{demog_change_component_df}.}

\item{maxsum}{integer, indicating how many levels should be shown for
    \code{\link[base]{factor}}s.}

\item{digits}{integer, used for number formatting with
    \code{\link[base]{signif}()} (for \code{summary.default}) or
    \code{\link[base]{format}()} (for \code{summary.data.frame}).  In
    \code{summary.default}, if not specified (i.e.,
    \code{\link[base]{missing}(.)}), \code{signif()} will \emph{not} be called
    anymore (since \R >= 3.4.0, where the default has been changed to
    only round in the \code{print} and \code{format} methods).}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list with elements \code{time}, \code{age}, \code{sex}, and \code{table}.
}
\description{

}
\author{
Mark Wheldon
}
