% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.vital_rate_df_time_age_sex}
\alias{summary.vital_rate_df_time_age_sex}
\title{Summarize a \code{vital_rate_df_time_age_sex}}
\usage{
\method{summary}{vital_rate_df_time_age_sex}(object, maxsum = 7, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An object of class \code{vital_rate_df_time_age_sex}.}

\item{maxsum}{integer, indicating how many levels should be shown for
    \code{\link[base]{factor}}s.}

\item{digits}{integer, used for number formatting with
    \code{\link[base]{signif}()} (for \code{summary.default}) or
    \code{\link[base]{format}()} (for \code{summary.data.frame}).  In
    \code{summary.default}, if not specified (i.e.,
    \code{\link[base]{missing}(.)}), \code{signif()} will \emph{not} be called
    anymore (since \R >= 3.4.0, where the default has been changed to
    only round in the \code{print} and \code{format} methods).}

\item{...}{additional arguments affecting the summary produced.}
}
\description{

}
\author{
Mark Wheldon
}
