% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_ccmpp_input_list.R
\name{summary_ccmpp_input_list}
\alias{summary_ccmpp_input_list}
\alias{summary.ccmpp_input_list}
\title{Summarize a \code{ccmpp_input_list}}
\usage{
\method{summary}{ccmpp_input_list}(object, vsep, ...)
}
\arguments{
\item{object}{An object of class \code{ccmpp_input_list}.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The object (invisibly)
}
\description{
A method for the generic \code{\link{summary}} function for objects
of class \code{\link{ccmpp_input_list}}. Summary statistics are returned invisibly in a named list.
}
\author{
Mark Wheldon
}
