% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sweden_1993}
\alias{sweden_1993}
\title{Example life table and population of Sweden, 1993}
\format{
A list with elements
\describe{
\item{P0F}{Female population by five-year age group in 1993.}
\item{P0M}{Male population by five-year age group in 1993.}
\item{LxF}{Female nLx values by five-year age group for the period [1993, 1998].}
\item{LxM}{Male nLx values by five-year age group for the period [1993, 1998].}
\item{Fx}{Age-specific fertility rates by five-year age group for the period [1993, 1998].}
\item{NMxF}{Age-specific net number of international female migrants for the period [1993, 1998].}
\item{NMxM}{Age-specific net number of international male migrants for the period [1993, 1998].}
\item{SxF}{Female nSx values by five-year age group for the period [1993, 1998].}
\item{SxM}{Female nSx values by five-year age group for the period [1993, 1998].}
\item{asfr}{Age-specific fertility rates by five-year age group for the period [1993, 1998].}
\item{srb}{Sex-ratio at birth for the period [1993, 1998].}
\item{n}{Width of the age intervals, in this case it is 5.}
\item{PzF}{Projected female population by five-year age group in 1998.}
\item{PzM}{Projected male population by five-year age group in 1998.}
}
}
\source{
Preston, S. H., Heuveline, P., and Guillot, M. (2001), \emph{Demography: Measuring and Modeling Population Processes}, Malden, Massachusetts: Blackwell.
}
\usage{
sweden_1993
}
\description{
An example life table and population counts of the population of
Sweden in 1993, based on the female life table and counts in
\cite{Preston et. al, 2001, Ch. 6, Box 6.2}.
}
\keyword{datasets}
