% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sweden_1993_Preston_5x5}
\alias{sweden_1993_Preston_5x5}
\title{Example abridged life table and population of Sweden, 1993}
\format{
A list with elements
\describe{
\item{pop_count_age_f_begin}{Female population by five-year age group in 1993.}
\item{pop_count_age_m_begin}{Male population by five-year age group in 1993.}
\item{pers_years_age_f}{Female nLx values by five-year age group for the period [1993, 1998].}
\item{pers_years_age_m}{Male nLx values by five-year age group for the period [1993, 1998].}
\item{fert_rate_age_f}{Age-specific fertility rates by five-year age group for the period [1993, 1998].}
\item{mig_net_count_age_f}{Age-specific net number of international female migrants for the period [1993, 1998].}
\item{mig_net_count_age_m}{Age-specific net number of international male migrants for the period [1993, 1998].}
\item{survival_ratio_age_f}{Female nSx values by five-year age group for the period [1993, 1998].}
\item{survival_ratio_age_m}{Female nSx values by five-year age group for the period [1993, 1998].}
\item{srb}{Sex-ratio at birth for the period [1993, 1998].}
\item{age_width}{Width of the age intervals, in this case it is 5.}
\item{birth_count_age_b}{births both sexes by five-year age group of the mother for the period [1993, 1998].}
\item{pop_count_age_f_end}{Projected female population by five-year age group in 1998.}
\item{pop_count_age_m_end}{Projected male population by five-year age group in 1998.}
}
}
\source{
Preston, S. H., Heuveline, P., and Guillot, M. (2001), \emph{Demography: Measuring and Modeling Population Processes}, Malden, Massachusetts: Blackwell.
}
\usage{
sweden_1993_Preston_5x5
}
\description{
An example life table and population counts of the population of
Sweden in 1993, based on the female life table and counts in
\cite{Preston et. al, 2001, Ch. 6, Box 6.2}.
}
\keyword{datasets}
