% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_vital_rate_df_access_attr.R
\name{test_vital_rate_dimensions}
\alias{test_vital_rate_dimensions}
\alias{is_by_time}
\alias{is_by_time.vital_rate_df}
\alias{is_by_time.data.frame}
\alias{is_by_age}
\alias{is_by_age.vital_rate_df}
\alias{is_by_age.data.frame}
\alias{is_by_sex}
\alias{is_by_sex.vital_rate_df}
\alias{is_by_sex.data.frame}
\title{Test for vital rate dimensions}
\usage{
is_by_time(x)

\method{is_by_time}{vital_rate_df}(x)

\method{is_by_time}{data.frame}(x)

is_by_age(x)

\method{is_by_age}{vital_rate_df}(x)

\method{is_by_age}{data.frame}(x)

is_by_sex(x)

\method{is_by_sex}{vital_rate_df}(x)

\method{is_by_sex}{data.frame}(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} or \code{FALSE} depending on the result.
}
\description{
These functions test whether an object has a particular vital rate
dimension. For objects of class \code{vital_rate_df} this uses the
attributes. For data frames it merely checks the presence of the
appropriate column.
}
\author{
Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon
}
