% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{validate_demog_change_component_df}
\alias{validate_demog_change_component_df}
\title{Validate objects of class \code{demog_change_component_df}.}
\usage{
validate_demog_change_component_df(x)
}
\arguments{
\item{x}{An object of class \code{vital_rate_data_frame}.}
}
\value{
Either an error or the object \code{x}.
}
\description{
Checks that an object with \code{class} attribute
\code{vital_rate_data_frame} is a valid object of this type.
}
\seealso{
vital_rate_data_frame
}
\author{
Mark Wheldon
}
\concept{vital_rate_data_frame class non-exported functions}
