% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_value_scale.R
\name{value_scale}
\alias{value_scale}
\alias{value_scale.demog_change_component_df}
\alias{value_scale_numeric}
\alias{value_scale_numeric.demog_change_component_df}
\alias{print.demog_change_component_df_value_scale}
\alias{value_scale<-}
\alias{value_scale<-.demog_change_component_df}
\alias{value_scale.ccmpp_input_df}
\title{Create, extract and set a \dQuote{value_scale}}
\usage{
value_scale(x)

\method{value_scale}{demog_change_component_df}(x)

value_scale_numeric(x)

\method{value_scale_numeric}{demog_change_component_df}(x)

\method{print}{demog_change_component_df_value_scale}(x, ...)

value_scale(x, ...) <- value

\method{value_scale}{demog_change_component_df}(x, ...) <- value

\method{value_scale}{ccmpp_input_df}(x)
}
\arguments{
\item{x}{An object inheriting from
\code{demog_change_component_df}.}
}
\value{
For \code{value_scale}, an object of class
    \code{\var{[prefix]}_value_scale}. For
    \code{value_scale_numeric} a numeric scalar with no class. See
    \dQuote{Details}.
}
\description{
These functions extract and set the \dQuote{value_scale} attributes
of objects inheriting from
\code{\link{demog_change_component_df}}. This attribute is
displayed differently depending on the class and other attributes
of its parent object. If you are comparing \dQuote{value_scales}
with, e.g., \code{\link{identical}} you should use
\code{value_scale_numeric} (see \dQuote{Details}).
}
\details{
The object returned by \code{value_scale} is given class
\code{\var{[prefix]}_value_scale}. \var{[prefix]} can currently be
either \code{demog_change_component_df} or \code{ccmpp_input_df}
depending on the class of the input \code{x}. This is currently the
only (exported) way of creating objects of these types; the class
is used only to control the \code{\print} method. As a consequence,
for example, \code{attr(x, "value_scale")} will return the
unclassed, numeric, value, not an object of class
\code{\var{[prefix]}_value_scale} (but see function
\code{value_scale_numeric} below).

Function \code{value_scale_numeric} returns a numeric (double) scalar (via
\code{\link{as.double}}) with no class attribute. It is suitable
for use with \code{\link{identical}} when comparing the
\dQuote{value_scale}s of different objects.
}
\seealso{
Other extract_attributes: 
\code{\link{age_span}()},
\code{\link{extract_ccmpp_input_list_attributes}}
}
\author{
Mark Wheldon
}
\concept{extract_attributes}
