% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{vital_rate_attributes}
\alias{vital_rate_attributes}
\title{Extract attributes specific to \code{vital_rate_df}s}
\usage{
vital_rate_attributes(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
A list of attributes.
}
\description{
All attributes that are not attributes of \code{data.frame} objects
are extracted. This function is designed to work with objects of
class \code{vital_rate_df}; behaviour for other classes is
not defined.
}
\author{
Mark Wheldon
}
