% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_vital_rate_df_access_attr.R
\name{vital_rate_dimensions}
\alias{vital_rate_dimensions}
\title{Extract the \dQuote{dimensions} attribute}
\usage{
vital_rate_dimensions(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
The extracted attribute
}
\description{
Extracts the non-data.frame attributes. The only method that exists
is for \code{vital_rate_df} objects.
}
\author{
Mark Wheldon
}
