% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpp_input_example}
\alias{wpp_input_example}
\title{Example inputs file for WPP 1x1 workflow, one country and variant}
\format{
A list with elements
\describe{
\item{pop_count_age_sex_base}{A data frame with population counts by age and sex in the base year.}
\item{life_table_age_sex}{A data frame of life table values by single year of age and sex by year from 1950 to 2019.}
\item{fert_rate_age_f}{A data frame of age-specific fertility rates by single year of age by year from 1950 to 2019.}
\item{srb}{A data frame of the sex ratio at birth (males/females) by year from 1950-2019.}
\item{mig_net_count_age_sex}{A data frame of net migrant counts by single year of age and sex by year from 1950-2019.}
\item{mig_net_rate_age_sex}{A data frame of net migrant rates by single year of age and sex by year from 1950-2019.}
\item{mig_net_count_tot_b}{A data frame of total net migrant counts (both sexes) year from 1950-2019.}
\item{mig_parameter}{A data frame with indicators of what type of net migration input to use ("counts", "rates", or "totals") and
whether migration should be accounted at "end" of period or "even" over period by year from 1950-2019.}
}
}
\source{
Interpolated from estimates for Canada published in United Nations, \emph{World Population Prospects 2019}.
}
\usage{
wpp_input_example
}
\description{
All inputs needed to perform a cohort component population projection
by single year of age and in one-year steps for the projection (1x1) from 1950-2020
}
\keyword{datasets}
