% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{aggregate.demog_change_component_df}
\alias{aggregate.demog_change_component_df}
\title{Aggregate a \code{demog_change_component_df}}
\usage{
\method{aggregate}{demog_change_component_df}(x, by, FUN = "sum", by_type = c("dimensions", "columns"), ...)
}
\arguments{
\item{x}{An object inheriting from
\code{demog_change_component_df}.}

\item{by}{A \emph{character vector} (not a list) of
\dQuote{dimensions} to aggregate by (or columns if
\code{by_type} is \code{\dQuote{columns}}; see
\dQuote{Details}).}

\item{FUN}{A function to use to aggregate the \dQuote{value} column
of \code{x}.}

\item{by_type}{Character argument specifying whether \code{by}
provides \dQuote{dimensions} or column names to aggregate by.}

\item{...}{Passed to \code{\link[stats]{aggregate}}.}
}
\value{
A data frame with the column \dQuote{value} after
    aggregation by \code{by} (see \dQuote{Description}).
}
\description{
This is a simple method for the generic
\code{\link[stats]{aggregate}} function in the \pkg{stats} package
for objects inheriting from \code{demog_change_component_df}. The
\dQuote{value} column is aggregated using function \code{FUN}; you
cannot aggregate any other columns of \code{x} with this
method. Note that argument \code{by} is a character vector, not a
list (see \dQuote{Details} for more information). The \code{by}
variables are the ones that will appear in the output, not the ones
that are collapsed.
}
\details{
Argument \code{by} is a character vector (unlike the default
method) and can be any of the allowed \dQuote{dimensions} (default)
\emph{or} column names of \code{x}. You must specify \code{by_type
= "columns"} if you are supplying column names.

The function will try to return an object of class
\code{demog_change_component_df} (regardless of the class of
\code{x}). If a valid object of this class cannot be created from
the result, a \code{\link{data.frame}} is returned (and with a
warning in interactive sessions).

This method is a rather simple wrapper for conveniently aggregating
\code{demog_change_component_df} objects by their dimensions (or
column names if desired). After processing the \code{by} and
\code{by_type} arguments, the default method for \code{aggregate}
is called (which, in turn, will call the \code{data.frame} method).

The \dQuote{time_span} and
\dQuote{age_span} columns, if present in \code{x}, will not be
present in the result (unless \code{by_type = "columns"} and they
were included in \code{by}). Pass the output back into
\code{\link{demog_change_component_df}} to check for validity and
recompute the \dQuote{span} columns, if required.
}
\seealso{
\code{\link[stats]{aggregate}}
    \code{\link{demog_change_component_df}}
}
\author{
Mark Wheldon
}
