% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{aggregate.demog_change_component_df}
\alias{aggregate.demog_change_component_df}
\title{Aggregate a \code{demog_change_component_df}}
\usage{
\method{aggregate}{demog_change_component_df}(x, by, FUN, ..., simplify = TRUE, drop = TRUE)
}
\description{
A method for the generic \code{\link[stats]{aggregate}}
function. This merely calls the
\code{\link[stats]{aggregate.data.frame}} method and issues a
warning that the class will be dropped (if called
interactively). See \code{link{collapse_demog_dimension}} for an
alternative if aggregation across an entire demographic dimension
(e.g., \dQuote{age}, \dQuote{time}, etc.) is desired.
}
\seealso{
\code{\link[stats]{aggregate}}, \code{link{collapse_demog_dimension}}
}
\author{
Mark Wheldon
}
