% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_age_time_matrix.R,
%   R/S3_classes_fert_rate_input_df.R
\name{as_age_time_matrix}
\alias{as_age_time_matrix}
\alias{as_age_time_matrix.demog_change_component_df}
\alias{as_age_time_matrix.fert_rate_age_f}
\title{Coerce a \code{demog_change_component_df} to \code{matrix}}
\usage{
as_age_time_matrix(x, ...)

\method{as_age_time_matrix}{demog_change_component_df}(x, ...)

\method{as_age_time_matrix}{fert_rate_age_f}(x, drop_zero_fert_ages = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from class
\code{demog_change_component_df}.}

\item{...}{Further arguments passed to and from methods.}

\item{drop_zero_fert_ages}{Logical; should rows corresponding to
\dQuote{non_zero_fert_ages} be dropped when converting
\code{fert_rate_age_f} objects.}
}
\value{
A \code{matrix}.
}
\description{
An age * time matrix will be created from the input object,
\code{x}, which must not have a \dQuote{sex} dimension. To create a
matrix from such an object, first remove the \dQuote{sex} dimension
with \code{\link{subset_sex}}.
}
\author{
Mark Wheldon
}
