% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df.R
\name{ccmpp_input_df}
\alias{ccmpp_input_df}
\title{Constructor for class \code{ccmpp_input_df}}
\usage{
ccmpp_input_df(
  x,
  dimensions = attr(x, "dimensions"),
  age_span = attr(x, "age_span"),
  time_span = attr(x, "time_span"),
  value_type = attr(x, "value_type"),
  ...
)
}
\arguments{
\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{ccmpp_input_df}.
}
\description{
**TBC** More strict version of \code{\link{demog_change_component_df}}.
}
\author{
Mark Wheldon
#' @export
}
