% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{ccmpp_input_list}
\alias{ccmpp_input_list}
\title{Constructor for class \code{ccmpp_input_list}}
\usage{
ccmpp_input_list(
  pop_count_base_input_df,
  life_table_input_df,
  fert_rate_input_df,
  srb_input_df,
  mig_net_count_input_df,
  mig_net_rate_input_df,
  mig_net_count_tot_input_df,
  mig_parameter_input_df
)
}
\arguments{
\item{pop_count_base_input_df}{An object of class \code{\link{pop_count_base_input_df}}}

\item{life_table_input_df}{An object of class \code{\link{life_table_input_df}}}

\item{fert_rate_input_df}{An object of class \code{\link{fert_rate_input_df}}}

\item{srb_input_df}{An object of class \code{\link{srb_input_df}}}

\item{mig_net_count_input_df}{An object of class \code{\link{mig_net_count_input_df}}}

\item{mig_net_rate_input_df}{An object of class \code{\link{mig_net_rate_input_df}}}

\item{mig_net_count_tot_input_df}{An object of class \code{\link{mig_net_count_tot_input_df}}}

\item{mig_parameter_input_df}{An object of class \code{\link{mig_parameter_input_df}}}
}
\value{
An object of class \code{ccmpp_input_list}.
}
\description{
This function creates objects of class \code{ccmpp_input_list}. It
takes as arguments a set of objects inheriting from
\code{\link{ccmpp_input_df}}.
}
\seealso{
\code{\link{as_ccmpp_input_list.list}} for creating
    \code{ccmpp_input_list}s lists.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_input_df}},
\code{\link{coerce_life_table_input_df}},
\code{\link{coerce_mig_net_count_input_df}},
\code{\link{coerce_mig_net_count_tot_input_df}},
\code{\link{coerce_mig_net_rate_input_df}},
\code{\link{coerce_mig_parameter_input_df}},
\code{\link{coerce_pop_count_base_input_df}},
\code{\link{coerce_srb_input_df}},
\code{\link{coerce_survival_ratio_input_df}},
\code{\link{fert_rate_input_df}()},
\code{\link{life_table_input_df}()},
\code{\link{mig_net_count_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{mig_parameter_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{srb_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
