% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list.R
\name{ccmpp_input_list}
\alias{ccmpp_input_list}
\title{Constructor for class \code{ccmpp_input_list}}
\usage{
ccmpp_input_list(
  pop_count_age_sex_base,
  life_table_age_sex,
  fert_rate_age_f,
  srb,
  mig_net_count_age_sex,
  mig_net_rate_age_sex,
  mig_net_count_tot_b,
  mig_parameter
)
}
\arguments{
\item{pop_count_age_sex_base}{An object of class \code{\link{pop_count_age_sex_base}}}

\item{life_table_age_sex}{An object of class \code{\link{life_table_age_sex}}}

\item{fert_rate_age_f}{An object of class \code{\link{fert_rate_age_f}}}

\item{srb}{An object of class \code{\link{srb}}}

\item{mig_net_count_age_sex}{An object of class \code{\link{mig_net_count_age_sex}}}

\item{mig_net_rate_age_sex}{An object of class \code{\link{mig_net_rate_age_sex}}}

\item{mig_net_count_tot_b}{An object of class \code{\link{mig_net_count_tot_b}}}

\item{mig_parameter}{An object of class \code{\link{mig_parameter}}}
}
\value{
An object of class \code{ccmpp_input_list}.
}
\description{
This function creates objects of class \code{ccmpp_input_list}. It
takes as arguments a set of objects inheriting from
\code{\link{ccmpp_input_df}}.
}
\seealso{
\code{\link{as_ccmpp_input_list.list}} for creating
    \code{ccmpp_input_list}s lists.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_input_df}},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
