% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_access_attr.R
\name{ccmpp_list_access_elements}
\alias{ccmpp_list_access_elements}
\alias{pop_count_base_component}
\alias{pop_count_base_component.ccmpp_input_list}
\alias{pop_count_base_component<-}
\alias{pop_count_base_component<-.ccmpp_input_list}
\alias{life_table_component}
\alias{life_table_component.ccmpp_input_list}
\alias{life_table_component<-}
\alias{life_table_component<-.ccmpp_input_list}
\alias{survival_ratio_component}
\alias{survival_ratio_component.life_table_age_sex}
\alias{survival_ratio_component.ccmpp_input_list}
\alias{survival_ratio_component<-}
\alias{survival_ratio_component<-.life_table_age_sex}
\alias{survival_ratio_component<-.ccmpp_input_list}
\alias{mortality_rate_component}
\alias{mortality_rate_component.life_table_age_sex}
\alias{mortality_rate_component.ccmpp_input_list}
\alias{mortality_rate_component<-}
\alias{mortality_rate_component<-.life_table_age_sex}
\alias{mortality_rate_component<-.ccmpp_input_list}
\alias{death_probability_component}
\alias{death_probability_component.life_table_age_sex}
\alias{death_probability_component.ccmpp_input_list}
\alias{death_probability_component<-}
\alias{death_probability_component<-.life_table_age_sex}
\alias{death_probability_component<-.ccmpp_input_list}
\alias{death_count_component}
\alias{death_count_component.life_table_age_sex}
\alias{death_count_component.ccmpp_input_list}
\alias{death_count_component<-}
\alias{death_count_component<-.life_table_age_sex}
\alias{death_count_component<-.ccmpp_input_list}
\alias{fert_rate_component}
\alias{fert_rate_component.ccmpp_input_list}
\alias{fert_rate_component<-}
\alias{fert_rate_component<-.ccmpp_input_list}
\alias{srb_component}
\alias{srb_component.ccmpp_input_list}
\alias{srb_component<-}
\alias{srb_component<-.ccmpp_input_list}
\alias{mig_net_count_component}
\alias{mig_net_count_component.ccmpp_input_list}
\alias{mig_net_count_component<-}
\alias{mig_net_count_component<-.ccmpp_input_list}
\alias{mig_net_rate_component}
\alias{mig_net_rate_component.ccmpp_input_list}
\alias{mig_net_rate_component<-}
\alias{mig_net_rate_component<-.ccmpp_input_list}
\alias{mig_net_count_tot_component}
\alias{mig_net_count_tot_component.ccmpp_input_list}
\alias{mig_net_count_tot_component<-}
\alias{mig_net_count_tot_component<-.ccmpp_input_list}
\alias{mig_parameter_component}
\alias{mig_parameter_component.ccmpp_input_list}
\alias{mig_parameter_component<-}
\alias{mig_parameter_component<-.ccmpp_input_list}
\title{Access the \code{ccmpp_input_df} components of a list}
\usage{
pop_count_base_component(x)

\method{pop_count_base_component}{ccmpp_input_list}(x)

pop_count_base_component(x) <- value

\method{pop_count_base_component}{ccmpp_input_list}(x) <- value

life_table_component(x)

\method{life_table_component}{ccmpp_input_list}(x)

life_table_component(x) <- value

\method{life_table_component}{ccmpp_input_list}(x) <- value

survival_ratio_component(x)

\method{survival_ratio_component}{life_table_age_sex}(x)

\method{survival_ratio_component}{ccmpp_input_list}(x)

survival_ratio_component(x) <- value

\method{survival_ratio_component}{life_table_age_sex}(x) <- value

\method{survival_ratio_component}{ccmpp_input_list}(x) <- value

mortality_rate_component(x)

\method{mortality_rate_component}{life_table_age_sex}(x)

\method{mortality_rate_component}{ccmpp_input_list}(x)

mortality_rate_component(x) <- value

\method{mortality_rate_component}{life_table_age_sex}(x) <- value

\method{mortality_rate_component}{ccmpp_input_list}(x) <- value

death_probability_component(x)

\method{death_probability_component}{life_table_age_sex}(x)

\method{death_probability_component}{ccmpp_input_list}(x)

death_probability_component(x) <- value

\method{death_probability_component}{life_table_age_sex}(x) <- value

\method{death_probability_component}{ccmpp_input_list}(x) <- value

death_count_component(x)

\method{death_count_component}{life_table_age_sex}(x)

\method{death_count_component}{ccmpp_input_list}(x)

death_count_component(x) <- value

\method{death_count_component}{life_table_age_sex}(x) <- value

\method{death_count_component}{ccmpp_input_list}(x) <- value

fert_rate_component(x)

\method{fert_rate_component}{ccmpp_input_list}(x)

fert_rate_component(x) <- value

\method{fert_rate_component}{ccmpp_input_list}(x) <- value

srb_component(x)

\method{srb_component}{ccmpp_input_list}(x)

srb_component(x) <- value

\method{srb_component}{ccmpp_input_list}(x) <- value

mig_net_count_component(x)

\method{mig_net_count_component}{ccmpp_input_list}(x)

mig_net_count_component(x) <- value

\method{mig_net_count_component}{ccmpp_input_list}(x) <- value

mig_net_rate_component(x)

\method{mig_net_rate_component}{ccmpp_input_list}(x)

mig_net_rate_component(x) <- value

\method{mig_net_rate_component}{ccmpp_input_list}(x) <- value

mig_net_count_tot_component(x)

\method{mig_net_count_tot_component}{ccmpp_input_list}(x)

mig_net_count_tot_component(x) <- value

\method{mig_net_count_tot_component}{ccmpp_input_list}(x) <- value

mig_parameter_component(x)

\method{mig_parameter_component}{ccmpp_input_list}(x)

mig_parameter_component(x) <- value

\method{mig_parameter_component}{ccmpp_input_list}(x) <- value
}
\arguments{
\item{x}{An object of class \code{ccmpp_input_list}}
}
\value{
An object inheriting from \code{ccmpp_input_df}.
}
\description{
These functions extract the \code{ccmpp_input_df} elements from an object
inheriting from \code{list}.
}
\author{
Mark Wheldon
}
