% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmppWPP_workflow_one_country_variant.R
\name{ccmppWPP_workflow_one_country_variant}
\alias{ccmppWPP_workflow_one_country_variant}
\title{Full ccmppWPP implementation for one country and variant}
\usage{
ccmppWPP_workflow_one_country_variant(wpp_input, intermediate_output_folder)
}
\arguments{
\item{wpp_input}{list of input objects required for one country-variant}
}
\value{
a list of objects including projected population, deaths, exposures, life tables, births, fertility
rates and proportions, and net migrant counts, needed for Eagle display
}
\description{
This function takes the wpp input list of objects for a given country and variant
and walks through the entire ccmppWPP workflow to produce all of the outputs published in the WPP.
}
\details{
This function calls all of the various functions required to complete the WPP workflow.  It currently
works only for a 1x1 population projection and returns outputs summarised by 1-year and 5-year age groups.
}
\author{
Sara Hertog
}
