% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df.R
\name{coerce_ccmpp_input_df}
\alias{coerce_ccmpp_input_df}
\alias{as_ccmpp_input_df}
\alias{as_ccmpp_input_df.data.frame}
\alias{as_ccmpp_input_df.matrix}
\alias{is_ccmpp_input_df}
\title{Coerce to a \code{ccmpp_input_df}}
\usage{
as_ccmpp_input_df(x, ...)

\method{as_ccmpp_input_df}{data.frame}(x, ...)

\method{as_ccmpp_input_df}{matrix}(x, ...)

is_ccmpp_input_df(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{ccmpp_input_df} if possible, or check if it is
one.
}
\section{Note on Validation}{

The method for \code{ccmpp_input_df} (i.e.,
\code{as_ccmpp_input_df.ccmpp_input_df}) does
\emph{not} check for validity. If \code{x} is somehow an invalid
\code{ccmpp_input_df} object, applying
\code{as_ccmpp_input_df} will simply return it (with
\code{attr(x, "class")} potentially modifed), not signal an
error. Use \code{ccmpp_input_df} on the object to ensure
validity is checked.
}

\author{
Mark Wheldon
}
