% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_net_count_tot_input_df.R
\name{coerce_mig_net_count_tot_b}
\alias{coerce_mig_net_count_tot_b}
\alias{as_mig_net_count_tot_b}
\alias{as_mig_net_count_tot_b.default}
\alias{as_mig_net_count_tot_b.data.frame}
\alias{as_mig_net_count_tot_b.matrix}
\alias{as_mig_net_count_tot_b.mig_net_count_tot_b}
\alias{is_mig_net_count_tot_b}
\title{Coerce to a \code{mig_net_count_tot_b}}
\usage{
as_mig_net_count_tot_b(x, ...)

\method{as_mig_net_count_tot_b}{default}(x, ...)

\method{as_mig_net_count_tot_b}{data.frame}(x, ...)

\method{as_mig_net_count_tot_b}{matrix}(x, ...)

\method{as_mig_net_count_tot_b}{mig_net_count_tot_b}(x, ...)

is_mig_net_count_tot_b(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{mig_net_count_tot_b} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
