% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_pop_count_base_input.R
\name{coerce_pop_count_age_sex_base}
\alias{coerce_pop_count_age_sex_base}
\alias{as_pop_count_age_sex_base}
\alias{as_pop_count_age_sex_base.data.frame}
\alias{as_pop_count_age_sex_base.matrix}
\alias{is_pop_count_age_sex_base}
\title{Coerce to a \code{pop_count_age_sex_base}}
\usage{
as_pop_count_age_sex_base(x, ...)

\method{as_pop_count_age_sex_base}{data.frame}(x, ...)

\method{as_pop_count_age_sex_base}{matrix}(x, ...)

is_pop_count_age_sex_base(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{pop_count_age_sex_base} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_pop_count_age_sex}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{pop_count_age_sex}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
