% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_srb_df.R
\name{coerce_srb}
\alias{coerce_srb}
\alias{as_srb}
\alias{as_srb.default}
\alias{as_srb.data.frame}
\alias{as_srb.matrix}
\alias{as_srb.srb}
\alias{is_srb}
\title{Coerce to a \code{srb}}
\usage{
as_srb(x, ...)

\method{as_srb}{default}(x, ...)

\method{as_srb}{data.frame}(x, ...)

\method{as_srb}{matrix}(x, ...)

\method{as_srb}{srb}(x, ...)

is_srb(x)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
    functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a
\code{srb} if possible, or check if it is
one.
}
\seealso{
\code{\link{coerce_demog_change_component_df}}

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mort_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_pop_count_age_sex_reference}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{death_probability_age_sex}()},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{mort_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
