% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{collapse_demog_dimension}
\alias{collapse_demog_dimension}
\title{Collapse a \code{demog_change_component_df} object across demographic dimension(s)}
\usage{
collapse_demog_dimension(
  x,
  FUN = "sum",
  ...,
  by_dimensions = demog_change_component_dims(x),
  collapse_dimensions = NULL,
  out_class = c("demog_change_component_df", "data.frame")
)
}
\arguments{
\item{x}{An object inheriting from
\code{demog_change_component_df}.}

\item{FUN}{A function to use to aggregate the \dQuote{value} column
of \code{x}; found by \code{\link{match.fun}}.}

\item{...}{Passed to
\code{\link[stats]{aggregate.data.frame}}. Must not include
arguments named \code{X}, \code{by}, or \code{FUN}.}

\item{by_dimensions}{A \emph{character vector} (not a list) of
demographic \dQuote{dimensions} to aggregate over; see
\dQuote{Details}).}

\item{out_class}{The class of the object to be returned
(specifically the first element of the vector returned by
\code{\link{class}}). See \dQuote{Setting the class of the output}.}
}
\value{
A data frame.
}
\description{
This function conveniently collapses objects across the demographic
dimension(s) specified in \code{by_dimensions}. The \dQuote{value}
column is aggregated using function \code{FUN}. Columns
corresponding to \code{by_dimensions} are dropped. If you want to
aggregate other columns, or complete collapsing over a dimension is
not desired, use \code{\link[stats]{aggregate}}, or similar.
}
\details{
Argument \code{by_dimensions} is a character vector and can be any
of the allowed demographic \dQuote{dimensions} such as
\dQuote{age}, \dQuote{time}, \dQuote{sex} (you can see the full
list by calling \code{ccmppWPP:::get_all_allowed_dimensions}). The
result will have values completely collapsed over the given
dimension. E.g., if \code{by_dimensions = "age"} then column
\code{value} in the result will contain the original \code{value}s
aggregated over all remaining dimensions, where aggregation is done
by function \code{FUN}.
}
\section{Setting the class of the output}{

Collapsing a \code{demog_change_component_df} object will likely
make it an invalid member of its class. By default, the function
will attempt to return a
\code{\link{demog_change_component_df}}. If this fails a
\code{\link{data.frame}} will be returned with a warning.
}

\seealso{
\code{\link[stats]{aggregate}}, \code{link{abridge}}
    \code{\link{demog_change_component_df}}
}
\author{
Mark Wheldon
}
