% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_access_attr.R
\name{extract_ccmpp_input_list_attributes}
\alias{extract_ccmpp_input_list_attributes}
\alias{age_span.ccmpp_input_list}
\alias{time_span.ccmpp_input_list}
\alias{ages.ccmpp_input_list}
\alias{non_zero_fert_ages.ccmpp_input_list}
\alias{non_zero_fert_ages<-.ccmpp_input_list}
\alias{times.ccmpp_input_list}
\alias{sexes.ccmpp_input_list}
\alias{indicators.ccmpp_input_list}
\alias{value_type.ccmpp_input_list}
\alias{value_scale.ccmpp_input_list}
\title{Extract attributes of a \code{ccmpp_input_list}}
\usage{
\method{age_span}{ccmpp_input_list}(x)

\method{time_span}{ccmpp_input_list}(x)

\method{ages}{ccmpp_input_list}(x)

\method{non_zero_fert_ages}{ccmpp_input_list}(x)

\method{non_zero_fert_ages}{ccmpp_input_list}(x) <- value

\method{times}{ccmpp_input_list}(x)

\method{sexes}{ccmpp_input_list}(x)

\method{indicators}{ccmpp_input_list}(x, drop = TRUE)

\method{value_type}{ccmpp_input_list}(x)

\method{value_scale}{ccmpp_input_list}(x)
}
\arguments{
\item{x}{An object inheriting from \code{ccmpp_input_list}}

\item{drop}{Logical, for \code{indicators}: should \code{NULL} elements be dropped?}
}
\value{
The requested attribute or dimension levels; see \dQuote{Details}.
}
\description{
These are convenience functions for accessing specific attributes
and dimensions of objects inheriting from
\code{ccmpp_input_list}.
}
\details{
\code{age_span} and \code{time_span} return the attribute, which
are scalars. \code{ages}, \code{times}, \code{sexes} return the
levels of these dimensions (see also
\code{\link{ages.demog_change_component_df}}). \code{indicators}
and \code{value_type} return lists, with one element per element of
\code{x}. By default, elements that do not have an
\dQuote{indicator} dimension are dropped by \code{indicators}.
}
\seealso{
Other extract_attributes: 
\code{\link{age_span}()}
}
\author{
Mark Wheldon
}
\concept{extract_attributes}
