% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_fert_rate_df.R,
%   R/S3_classes_demog_change_component_df_access_attr.R,
%   R/S3_classes_subclasses_access_attr.R
\name{non_zero_fert_ages}
\alias{non_zero_fert_ages}
\alias{non_zero_fert_ages.fert_rate_age_f}
\alias{non_zero_fert_ages<-}
\alias{non_zero_fert_ages<-.fert_rate_age_f}
\alias{extract_demog_change_component_attributes}
\alias{values}
\alias{values.demog_change_component_df}
\alias{values<-}
\alias{values<-.demog_change_component_df}
\alias{ages}
\alias{ages.demog_change_component_df}
\alias{times}
\alias{times.demog_change_component_df}
\alias{sexes}
\alias{sexes.demog_change_component_df}
\alias{indicators}
\alias{indicators.demog_change_component_df}
\alias{value_type}
\alias{value_type.demog_change_component_df}
\alias{value_type<-}
\alias{value_type<-.demog_change_component_df}
\alias{age_span}
\alias{age_span.ccmpp_input_df}
\alias{age_span.ccmpp_output_df}
\alias{time_span}
\alias{time_span.ccmpp_input_df}
\alias{time_span.ccmpp_output_df}
\alias{time_span.pop_count_age_sex_reference}
\alias{value_type<-.ccmpp_input_df}
\alias{value_type<-.ccmpp_output_df}
\alias{print.ccmpp_input_df_value_scale}
\title{Extract specific components}
\usage{
non_zero_fert_ages(x)

\method{non_zero_fert_ages}{fert_rate_age_f}(x)

non_zero_fert_ages(x, ...) <- value

\method{non_zero_fert_ages}{fert_rate_age_f}(x, ...) <- value

values(x)

\method{values}{demog_change_component_df}(x)

values(x) <- value

\method{values}{demog_change_component_df}(x) <- value

ages(x)

\method{ages}{demog_change_component_df}(x)

times(x)

\method{times}{demog_change_component_df}(x)

sexes(x)

\method{sexes}{demog_change_component_df}(x)

indicators(x)

\method{indicators}{demog_change_component_df}(x)

value_type(x)

\method{value_type}{demog_change_component_df}(x)

value_type(x, ...) <- value

\method{value_type}{demog_change_component_df}(x, ...) <- value

age_span(x)

\method{age_span}{ccmpp_input_df}(x)

\method{age_span}{ccmpp_output_df}(x)

time_span(x)

\method{time_span}{ccmpp_input_df}(x)

\method{time_span}{ccmpp_output_df}(x)

\method{time_span}{pop_count_age_sex_reference}(x)

\method{value_type}{ccmpp_input_df}(x, ...) <- value

\method{value_type}{ccmpp_output_df}(x, ...) <- value

\method{print}{ccmpp_input_df_value_scale}(x, ...)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
The \dQuote{levels} of the extracted component
    (except possibly for \code{values}; see \dQuote{Details}).
}
\description{
These functions extract specific components (attributes, factor
levels, values, etc.) from objects inheriting from
\code{demog_change_component_df}s. They return the
\emph{\dQuote{levels}} of the component, not the whole column from
the data frame. If the whole column is required use the usual
subsetting functions such as \code{\link{base::[}}. See
\dQuote{Extracting and replacing values} for the operation of
\code{values} and \code{values<-}, which are a bit different.
}
\section{Extracting and replacing values}{

If \code{\link{value_type(x)}} is \emph{not} \code{"categorical"}
then \code{value(x)} will return \code{x$value} as it is assumed
all are unique. The replacement function \code{values<-} replaces
the \code{"value"} column, using the usual recycling rules (see
\code{\link{data.frame}}), and attempts to return an object the
same class as \code{x}. If the replacement renders the object
invalid as a member of the class an error will be signaled.
}

\seealso{
Other extract_attributes: 
\code{\link{extract_ccmpp_input_list_attributes}},
\code{\link{value_scale}()}
}
\author{
Mark Wheldon
}
\concept{extract_attributes}
