% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_df_access_attr.R
\name{extract_demog_change_component_attributes}
\alias{extract_demog_change_component_attributes}
\alias{age_span}
\alias{time_span}
\alias{sexes}
\title{Extract specific vital rate attributes}
\usage{
age_span(x)

time_span(x)

sexes(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
The extracted attribute.
}
\description{
These functions extract specific attributes of
\code{demog_change_component_df}s. Methods for other classes are not defined.
}
\author{
Mark Wheldon

Mark Wheldon

Mark Wheldon

Mark Wheldon
}
