% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_df_access_attr.R
\name{extract_demog_change_component_attributes}
\alias{extract_demog_change_component_attributes}
\alias{age_span}
\alias{age_span.demog_change_component_df}
\alias{ages}
\alias{ages.demog_change_component_df}
\alias{time_span}
\alias{time_span.demog_change_component_df}
\alias{times}
\alias{times.demog_change_component_df}
\alias{sexes}
\alias{sexes.demog_change_component_df}
\alias{value_type}
\alias{value_type.demog_change_component_df}
\alias{non_zero_fert_ages}
\alias{non_zero_fert_ages.fert_rate_input_df}
\title{Extract specific attributes}
\usage{
age_span(x)

\method{age_span}{demog_change_component_df}(x)

ages(x)

\method{ages}{demog_change_component_df}(x)

time_span(x)

\method{time_span}{demog_change_component_df}(x)

times(x)

\method{times}{demog_change_component_df}(x)

sexes(x)

\method{sexes}{demog_change_component_df}(x)

value_type(x)

\method{value_type}{demog_change_component_df}(x)

non_zero_fert_ages(x)

\method{non_zero_fert_ages}{fert_rate_input_df}(x)
}
\arguments{
\item{x}{An object from which to extract attributes.}
}
\value{
The extracted attribute.
}
\description{
These functions extract specific attributes of
\code{demog_change_component_df}s. Methods for other classes are not defined.
}
\author{
Mark Wheldon
}
