% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_fert_rate_df.R
\name{fert_rate_age_f}
\alias{fert_rate_age_f}
\alias{fert_rate_age_f.data.frame}
\alias{fert_rate_age_f.ccmpp_input_list}
\alias{fert_rate_age_f<-}
\title{Constructor for class \code{fert_rate_age_f}}
\usage{
fert_rate_age_f(x, ...)

\method{fert_rate_age_f}{data.frame}(
  x,
  non_zero_fert_ages = attr(x, "non_zero_fert_ages"),
  value_scale = attr(x, "value_scale")
)

\method{fert_rate_age_f}{ccmpp_input_list}(x)

fert_rate_age_f(x) <- value
}
\arguments{
\item{x}{An object for which a method is defined (see \dQuote{Details}).}

\item{...}{Passed to the low-level constructor.}

\item{non_zero_fert_ages}{Numeric vector of unique ages indicating
the reproductive age range. See the \dQuote{Details} section.}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{fert_rate_age_f}.
}
\description{
\code{fert_rate_age_f} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes four additional conditions:
\enumerate{
  \item{The \code{value_type} attribute is \dQuote{rate}.}
  \item{All elements of the \dQuote{value} column must be non-negative.}
  \item{There can be no 'sex' dimension; all fertility rate inputs
  must be female fertility rates.}
  \item{An additional attribute \code{non_zero_fert_ages}, which
  specifies the reproductive age range as a vector of
  \code{age_start} values (see \dQuote{Details}).}}
}
\details{
Methods are defined for \code{\link{data.frame}}s and
\code{\link{ccmpp_input_list}}s, and possibly other objects as
well. The \code{data.frame} method \dQuote{constructs} an object
from \code{x}. The \code{ccmpp_input_list} method \dQuote{extracts}
an object from \code{x}. There is also a replacement function which
complements the extraction methods.

An attempt will be made to guess \code{non_zero_fert_ages} if they
are not explicitly specified using the \code{non_zero_fert_ages}
argument.

If \code{non_zero_fert_ages} are specified, they must be a subset
of \code{ages(x)} and be equally spaced with spacing equal to
\code{age_span(x)}.

Elements of the \code{value} column corresponding to ages not
listed in \code{non_zero_fert_ages} will be overwritten with zeros
(specifically \code{as.double(0)}s).
}
\seealso{
\code{\link{validate_ccmppWPP_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_death_count_age_sex}},
\code{\link{coerce_death_probability_age_sex}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_mort_rate_age_sex}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_pop_count_age_sex_reference}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{death_count_age_sex}()},
\code{\link{death_probability_age_sex}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{mig_parameter}()},
\code{\link{mort_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
