% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{fert_rate_input_df}
\alias{fert_rate_input_df}
\title{Constructor for class \code{fert_rate_input_df}}
\usage{
fert_rate_input_df(x, non_zero_fert_ages = attr(x, "non_zero_fert_ages"))
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{non_zero_fert_ages}{Numeric vector of unique ages indicating
the reproductive age range.}
}
\value{
An object of class \code{fert_rate_input_df}.
}
\description{
\code{fert_rate_input_df} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes three additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{There can be no 'sex' dimension; all fertility rate inputs
  must be female fertility rates.}
  \item{The additional attribute \code{non_zero_fert_ages}
  represents the reproductive age range as a vector of
  \code{age_start} values.}}
An attempt will be made to guess \code{non_zero_fert_ages} if not supplied.
}
\section{Non-zero fert ages}{

\code{non_zero_fert_ages} must be a subset of \code{ages(x)} and
equally spaced, with spacing equal to \code{age_span(x)}.
}

\seealso{
\code{\link{validate_ccmpp_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{life_table_input_df}()},
\code{\link{mig_net_count_input_df}()},
\code{\link{mig_net_count_tot_input_df}()},
\code{\link{mig_net_rate_input_df}()},
\code{\link{pop_count_base_input_df}()},
\code{\link{srb_input_df}()},
\code{\link{survival_ratio_input_df}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
