% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R,
%   R/S3_methods_fert_rate_input_df.R
\name{generic_coerce_demog_change_component_df}
\alias{generic_coerce_demog_change_component_df}
\alias{as.data.frame.demog_change_component_df}
\alias{as.data.frame.fert_rate_input_df}
\title{Coerce a \code{demog_change_component_df}.}
\usage{
\method{as.data.frame}{demog_change_component_df}(x, ...)

\method{as.data.frame}{fert_rate_input_df}(x, restore_columns = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}
}
\value{
A coerced object.
}
\description{
The class-specific attributes will be lost.
}
\author{
Mark Wheldon
}
