% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DemoData_tier1_post_process.R
\name{get_census_years}
\alias{get_census_years}
\title{Get years in which censuses were conducted}
\usage{
get_census_years(loc_id)
}
\value{
A \emph{character} vector of years.
}
\description{

}
\details{
Years in which censuses were conducted are provided in
\code{data("census_years")} for many countries. Ranges of the form
\code{"xxxx-yyyy"} (e.g., \code{"2000-2005"}) are possible if a
rolling population count was conducted over several years. France,
for example, has conducted rolling censuses.
}
\references{
United Nations (2021). Standard country or area codes for
    statistical use (M49). Department of Economic and Social
    Affairs, Statistics
    Division. https://unstats.un.org/unsd/methodology/m49/
    (retrieved 1 April 2021).
}
\seealso{
Other DDSQL_extract: 
\code{\link{DDextract_get_pop_count_age_sex_reference}()},
\code{\link{DDextract_get_pop_count_reference_times}()},
\code{\link{parse_pop_count_reference_times_ranges}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
