% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData_extract_post_process.R
\name{get_census_years}
\alias{get_census_years}
\title{Get years in which censuses were conducted}
\usage{
get_census_years(loc_id)
}
\value{
A \emph{character} vector of years.
}
\description{

}
\references{
United Nations (2021). Standard country or area codes for
    statistical use (M49). Department of Economic and Social
    Affairs, Statistics
    Division. https://unstats.un.org/unsd/methodology/m49/
    (retrieved 1 April 2021).
}
\seealso{
Other DDSQL_extract: 
\code{\link{get_pop_count_age_sex_reference}()},
\code{\link{get_pop_count_reference_times}()},
\code{\link{parse_pop_count_reference_times_ranges}()}
}
\author{
Mark Wheldon
}
\concept{DDSQL_extract}
