% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_list_access_attr.R,
%   R/S3_classes_mig_parameter_input_df.R
\name{mig_assumption.ccmpp_input_list}
\alias{mig_assumption.ccmpp_input_list}
\alias{mig_assumption<-.ccmpp_input_list}
\alias{mig_assumption_extract_and_set}
\alias{mig_assumption}
\alias{mig_assumption.mig_parameter}
\alias{mig_assumption<-}
\alias{mig_assumption<-.mig_parameter}
\title{Extract or set the migration assumption}
\usage{
\method{mig_assumption}{ccmpp_input_list}(x)

\method{mig_assumption}{ccmpp_input_list}(x) <- value

mig_assumption(x)

\method{mig_assumption}{mig_parameter}(x)

mig_assumption(x) <- value

\method{mig_assumption}{mig_parameter}(x) <- value
}
\arguments{
\item{x}{An object inheriting from \code{\link{mig_parameter}}.}
}
\value{
A character string with the assumption for the extraction
    function, or \code{x} (invisibly) with the assumption set to
    \code{value} for the replacement function.
}
\description{
Returns or sets the migration assumption for a
\code{\link{mig_parameter}} object. Migration assumption is stored
in the \value{column} in rows where \code{indicator} =
\dQuote{mig_assumption}.
}
\author{
Mark Wheldon
}
