% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_mig_parameter_input_df.R
\name{mig_parameter}
\alias{mig_parameter}
\title{Constructor for class \code{mig_parameter}}
\usage{
mig_parameter(x, time_span = attr(x, "time_span"))
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}
}
\value{
An object of class \code{mig_parameter}.
}
\description{
\code{mig_parameter} is a subclass of
\code{\link{ccmpp_input_df}}. It imposes two additional conditions:
\enumerate{
  \item{\code{Value_type} attribute equals \dQuote{rate}.}
  \item{Within year and sex, age must start at 0.}}
}
\seealso{
\code{\link{validate_ccmpp_object}} for object validation,
    \code{\link{ccmpp_input_df}} for the class from which this one
    inherits.

Other ccmpp_input_objects: 
\code{\link{ccmpp_input_df}()},
\code{\link{ccmpp_input_list}()},
\code{\link{coerce_ccmpp_input_df}},
\code{\link{coerce_ccmpp_input_list}},
\code{\link{coerce_demog_change_component_df}},
\code{\link{coerce_fert_rate_age_f}},
\code{\link{coerce_life_table_age_sex}},
\code{\link{coerce_mig_net_count_age_sex}},
\code{\link{coerce_mig_net_count_tot_b}},
\code{\link{coerce_mig_net_prop_age_sex}},
\code{\link{coerce_mig_net_rate_age_sex}},
\code{\link{coerce_mig_parameter}},
\code{\link{coerce_pop_count_age_sex_base}},
\code{\link{coerce_srb}},
\code{\link{coerce_survival_ratio_age_sex}},
\code{\link{fert_rate_age_f}()},
\code{\link{life_table_age_sex}()},
\code{\link{mig_net_count_age_sex}()},
\code{\link{mig_net_count_tot_b}()},
\code{\link{mig_net_prop_age_sex}()},
\code{\link{mig_net_rate_age_sex}()},
\code{\link{pop_count_age_sex_base}()},
\code{\link{srb}()},
\code{\link{survival_ratio_age_sex}()}
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_objects}
