% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_df.R
\name{new_ccmpp_input_df}
\alias{new_ccmpp_input_df}
\title{Low-level constructor for class \code{ccmpp_input_df}.}
\usage{
new_ccmpp_input_df(
  x,
  dimensions = character(),
  age_span = double(),
  time_span = double(),
  value_type = character(),
  ...,
  class = character()
)
}
\arguments{
\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}
}
\value{
An object of class \code{ccmpp_input_df}.
}
\description{
Creates an object of class \code{ccmpp_input_df}. Minimal
checks are done; for interactive use see
\code{\link{validate_ccmpp_input_df}}.

This function is not exported. The user-level constructor is
\code{\link{ccmpp_input_df}}.
}
\seealso{
ccmpp_input_df
}
\author{
Mark Wheldon
}
\concept{ccmpp_input_df class non-exported functions}
