% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{new_demog_change_component_df}
\alias{new_demog_change_component_df}
\title{Low-level constructor for class \code{demog_change_component_df}.}
\usage{
new_demog_change_component_df(
  x,
  dimensions = character(),
  value_type = character(),
  value_scale = double(),
  ...,
  class = character()
)
}
\arguments{
\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}

\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}
}
\value{
An object of class \code{demog_change_component_df}.
}
\description{
Creates an object of class \code{demog_change_component_df}. Minimal
checks are done; for interactive use see
\code{\link{demog_change_component_df}}.

This function is not exported. The user-level constructor is
\code{\link{demog_change_component_df}}.
}
\seealso{
demog_change_component_df
}
\author{
Mark Wheldon
}
