% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_fert_rate_input_df.R
\name{new_fert_rate_age_f}
\alias{new_fert_rate_age_f}
\title{Low-level constructor for class \code{fert_rate_age_f}.}
\usage{
new_fert_rate_age_f(
  x,
  age_span = double(),
  time_span = double(),
  dimensions = get_req_dimensions_for_ccmpp_input_classes("fert_rate_age_f"),
  value_type = get_value_types_for_classes("fert_rate_age_f"),
  value_scale = double(),
  non_zero_fert_ages = double(),
  ...,
  class = character()
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}

\item{...}{Passed to the low-level constructor.}
}
\value{
An object of class \code{fert_rate_age_f}.
}
\description{
Creates an object of class \code{fert_rate_age_f}. Minimal
checks are done; for interactive use see
\code{\link{fert_rate_age_f}}.

This function is not exported. The user-level constructor is
\code{\link{fert_rate_age_f}}.
}
\seealso{
fert_rate_age_f
}
\author{
Mark Wheldon
}
