% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_mig_net_count_tot_df.R
\name{new_mig_net_count_tot_b}
\alias{new_mig_net_count_tot_b}
\title{Low-level constructor for class \code{mig_net_count_tot_b}.}
\usage{
new_mig_net_count_tot_b(
  x,
  age_span = double(),
  time_span = double(),
  dimensions = get_req_dimensions_for_ccmpp_in_out_classes("mig_net_count_tot_b"),
  value_type = get_value_types_for_ccmpp_in_out_classes("mig_net_count_tot_b"),
  value_scale = double(),
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{mig_net_count_tot_b}.
}
\description{
Creates an object of class \code{mig_net_count_tot_b}. Minimal
checks are done; for interactive use see
\code{\link{mig_net_count_tot_b}}.

This function is not exported. The user-level constructor is
\code{\link{mig_net_count_tot_b}}.
}
\seealso{
mig_net_count_tot_b
}
\author{
Mark Wheldon
}
