% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_pop_count_base_input.R
\name{new_pop_count_age_sex_base}
\alias{new_pop_count_age_sex_base}
\title{Low-level constructor for class \code{pop_count_age_sex_base}.}
\usage{
new_pop_count_age_sex_base(
  x,
  age_span = double(),
  time_span = double(),
  dimensions = get_req_dimensions_for_ccmpp_in_out_classes("pop_count_age_sex_base"),
  value_type = get_value_types_for_ccmpp_in_out_classes("pop_count_age_sex_base"),
  value_scale = double(),
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{pop_count_age_sex_base}.
}
\description{
Creates an object of class \code{pop_count_age_sex_base}. Minimal
checks are done; for interactive use see
\code{\link{pop_count_age_sex_base}}.

This function is not exported. The user-level constructor is
\code{\link{pop_count_age_sex_base}}.
}
\seealso{
pop_count_age_sex_base
}
\author{
Mark Wheldon
}
\concept{pop_count_age_sex_base class non-exported functions}
