% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_srb_df.R
\name{new_srb}
\alias{new_srb}
\title{Low-level constructor for class \code{srb}.}
\usage{
new_srb(
  x,
  age_span = double(),
  time_span = double(),
  dimensions = get_req_dimensions_for_subclass_classes("srb"),
  value_type = get_value_types_for_subclass_classes("srb"),
  value_scale = NA,
  ...,
  class = character()
)
}
\arguments{
\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{srb}.
}
\description{
Creates an object of class \code{srb}. Minimal
checks are done; for interactive use see
\code{\link{srb}}.

This function is not exported. The user-level constructor is
\code{\link{srb}}.
}
\seealso{
srb
}
\author{
Mark Wheldon
}
