% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_survival_ratio_df.R
\name{new_survival_ratio_age_sex}
\alias{new_survival_ratio_age_sex}
\title{Low-level constructor for class \code{survival_ratio_age_sex}.}
\usage{
new_survival_ratio_age_sex(
  x,
  age_span = double(),
  time_span = double(),
  non_zero_fert_ages = double(),
  dimensions = get_req_dimensions_for_ccmpp_in_out_classes("survival_ratio_age_sex"),
  value_type = get_value_types_for_ccmpp_in_out_classes("survival_ratio_age_sex"),
  value_scale = NA,
  ...,
  class = character()
)
}
\arguments{
\item{age_span}{Scalar indicating the span of the age groups.}

\item{time_span}{Scalar indicating the span of the time periods.}

\item{dimensions}{Character vector listing the dimensions such as
\dQuote{time}, \dQuote{age}, \dQuote{sex}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
An object of class \code{survival_ratio_age_sex}.
}
\description{
Creates an object of class \code{survival_ratio_age_sex}. Minimal
checks are done; for interactive use see
\code{\link{survival_ratio_age_sex}}.

This function is not exported. The user-level constructor is
\code{\link{survival_ratio_age_sex}}.
}
\seealso{
survival_ratio_age_sex
}
\author{
Mark Wheldon
}
