% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_variants.R
\name{pasfr_given_tfr}
\alias{pasfr_given_tfr}
\title{Infer proportion age-specific fertility rates from tfr using a global normative model}
\usage{
pasfr_given_tfr(
  PasfrGlobalNorm,
  pasfr_observed,
  tfr_observed_projected,
  years_projection = 2020:2100,
  num_points = 15
)
}
\arguments{
\item{PasfrGlobalNorm}{matrix. output from pasfr_global_model() function}

\item{pasfr_observed}{matrix. with age in single years from 10 to 54 in rows and year in columns. Value is country estimates of pasfr}

\item{tfr_observed_projected}{numeric vector. TFR values for the country. Names are years.}

\item{years_projection}{numeric vector.  Years for which pasfr is to be projected}

\item{num_points}{numeric scalar. The number of past point estimates to be used to model pasfr}
}
\value{
a matrix of pasfr with age in rows and years in columns
}
\description{
This function returns the proportional age-specific fertility rates given a projected level of
total fertility by converging the initial pasfr to a global model
}
\details{
accesses the global model in the data frame "pasfr_global_model"
}
\author{
Sara Hertog
}
