% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_variants.R
\name{pasfr_global_model}
\alias{pasfr_global_model}
\title{Compute global normative model of proportion age-specific fertility rates}
\usage{
pasfr_global_model(
  tfr_all_locs = "data/tfr_estimates_projections_for_pasfr_global_norm.RData",
  pasfr_all_locs = "data/pasfr_estimates_for_pasfr_global_norm.RData",
  pasfr_pattern = "data/PASFRpattern.RData",
  present_year = 2019
)
}
\arguments{
\item{tfr_all_locs}{character. file path to RData file with annual TFR; countries in rows, years in columns}

\item{pasfr_all_locs}{character. file path to RData file with 1x1 PASFR; countries and ages from 10 to 54 in rows, years in columns}

\item{pasfr_pattern}{character. file path to RData file with 1 record per LocID; country_code = LocID, PASFRNorm = "Global Norm", 
PasfrGlobalNorm is c(0,1)flag indicating whether the countrys data should be used in computing global norm.}

\item{present_year}{numeric. Last year of observed data?}
}
\value{
a matrix of annual PASFR norms; age from 10 to 54 in row, year in columns
}
\description{
This function calls bayesPop functions to return the proportional age-specific fertility rates associated with a global norm
based on historical patterns of TFR and PASFR
}
\details{
calls embedded functions from bayesPop
}
\author{
Sara Hertog
}
