% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{plot.demog_change_component_df}
\alias{plot.demog_change_component_df}
\title{Plot \code{demog_change_component_df} objects}
\usage{
\method{plot}{demog_change_component_df}(
  x,
  type = c("point", "line", "both"),
  ...,
  framework = c("ggplot2", "base"),
  plot = TRUE
)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}

\item{type}{The type of plot to produce, e.g., \dQuote{line}.}

\item{framework}{Plotting framework to use; either \dQuote{ggplot2}
or \dQuote{base}. Use the former (default) for better results.}

\item{plot}{Should the plot be printed?}
}
\value{
When \code{framework} is \dQuote{ggplot2}, a plot
    object. Otherwise nothing. The function is mainly called for
    the side effect of producing a plot.
}
\description{
This is the S3 method for the generic \code{\link{plot}} function.
}
\author{
Mark Wheldon
}
