% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{prepare_df_for_demog_change_component_df}
\alias{prepare_df_for_demog_change_component_df}
\title{Prepare a data frame for conversion to a \code{demog_change_component_df}}
\usage{
prepare_df_for_demog_change_component_df(
  x,
  dimensions = attr(x, "dimensions"),
  value_type = attr(x, "value_type"),
  value_scale = attr(x, "value_scale")
)
}
\arguments{
\item{x}{For \code{demog_change_component_df}: A data frame with
columns \dQuote{age_start}, \dQuote{age_span}, \dQuote{sex},
\dQuote{time_start}, \dQuote{time_span}, and \dQuote{value}.}

\item{value_type}{Scalar indicating the type of the \dQuote{value}
column (e.g., \dQuote{count}, \dQuote{rate}, etc.).}

\item{value_scale}{\emph{Numeric} scalar indicating the value_scale of the
counts, e.g., \code{1}, \code{1000}, \code{1e6}, etc.}
}
\value{
A data frame
}
\description{
Converts and cleans \code{x} into a form suitable for
\code{\link{demog_change_component_df}}.
}
\author{
Mark Wheldon
}
