% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_demog_change_component_df.R
\name{print.demog_change_component_df}
\alias{print.demog_change_component_df}
\title{Print Values of a \code{demog_change_component_df}}
\usage{
\method{print}{demog_change_component_df}(
  x,
  ...,
  n = min(6L, nrow(x)),
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = FALSE,
  na.print = ".",
  print_what = c("info", "table")
)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}

\item{...}{optional arguments to \code{print} methods.}

\item{n}{Integer controlling how many rows of \code{x} are
printed. Passed to \code{link{head}}; see the documentation of
that function for valid values and associated behaviours.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{na.print}{character string (or \code{NULL}) indicating
    \code{\link[base]{NA}} values in printed output, see
    \code{\link[base]{print.default}}.}

\item{print_what}{What should be printed?}
}
\description{
This is a method for the generic \code{\link{base::print}}
function. Only the first \code{n} rows are printed for convenience
(by default). If all rows are desired use \code{as.data.frame(x)}
or see the definition of argument \code{n}.
}
\details{
Printed output consists
of some text, \code{"info"}, and the first \code{n} lines of the
data, \code{"table"}. What gets printed is controlled by argument
\code{print_what}.
}
\author{
Mark Wheldon
}
