% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S3_classes_demog_change_component_utility_functions.R
\name{subset_demog_change_component_df}
\alias{subset_demog_change_component_df}
\alias{subset_time}
\alias{subset_time.demog_change_component_df}
\alias{subset_age}
\alias{subset_age.demog_change_component_df}
\alias{subset_sex}
\alias{subset_sex.demog_change_component_df}
\title{Subset by time, age, or sex}
\usage{
subset_time(x, ...)

\method{subset_time}{demog_change_component_df}(x, time, drop = FALSE)

subset_age(x, ...)

\method{subset_age}{demog_change_component_df}(x, age, drop = FALSE)

subset_sex(x, ...)

\method{subset_sex}{demog_change_component_df}(x, sex = get_allowed_sexes(), drop = FALSE)
}
\arguments{
\item{x}{An object to subset.}

\item{time, age, sex}{Vectors indicating the levels of time, age, or
sex to retain.}

\item{drop}{Logical; should demographic change component dimensions
with only one level be dropped?}
}
\value{
The object after subsetting.
}
\description{
These functions subset \code{demog_change_component_df}s by one of
the three dimensions, age, time, or sex. Subsetting such that only
one level of the subset dimension is retained will drop the
dimension if \code{drop = TRUE}, otherwise it is retained (default).
}
\author{
Mark Wheldon
}
