% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_ccmpp_input_death_count_df.R,
%   R/S3_classes_ccmpp_input_death_probability_df.R,
%   R/S3_classes_ccmpp_input_df_utility_functions.R,
%   R/S3_classes_ccmpp_input_fert_rate_df.R,
%   R/S3_classes_ccmpp_input_life_table_df.R,
%   R/S3_classes_ccmpp_input_mig_net_count_df.R,
%   R/S3_classes_ccmpp_input_mig_net_count_tot_df.R,
%   R/S3_classes_ccmpp_input_mig_net_rate_df.R,
%   R/S3_classes_ccmpp_input_mig_parameter_df.R,
%   R/S3_classes_ccmpp_input_mort_rate_df.R,
%   R/S3_classes_ccmpp_input_pop_count_base_df.R,
%   R/S3_classes_ccmpp_input_srb_df.R,
%   R/S3_classes_ccmpp_input_survival_ratio_df.R,
%   R/S3_classes_ccmpp_ouput_df_utility_functions.R,
%   R/S3_classes_ccmpp_output_pop_count_age_sex_df.R,
%   R/S3_classes_demog_change_component_pop_count_reference.R,
%   R/S3_classes_demog_change_component_utility_functions.R
\name{subset_time.death_count_age_sex}
\alias{subset_time.death_count_age_sex}
\alias{subset_age.death_count_age_sex}
\alias{subset_sex.death_count_age_sex}
\alias{subset_time.death_probability_age_sex}
\alias{subset_age.death_probability_age_sex}
\alias{subset_sex.death_probability_age_sex}
\alias{subset_indicator.ccmpp_input_df}
\alias{subset_time.ccmpp_input_df}
\alias{subset_age.ccmpp_input_df}
\alias{subset_sex.ccmpp_input_df}
\alias{subset_time.fert_rate_age_f}
\alias{subset_age.fert_rate_age_f}
\alias{subset_indicator.life_table_age_sex}
\alias{subset_time.life_table_age_sex}
\alias{subset_age.life_table_age_sex}
\alias{subset_sex.life_table_age_sex}
\alias{subset_time.mig_net_count_age_sex}
\alias{subset_age.mig_net_count_age_sex}
\alias{subset_sex.mig_net_count_age_sex}
\alias{subset_time.mig_net_count_tot_b}
\alias{subset_time.mig_net_rate_age_sex}
\alias{subset_age.mig_net_rate_age_sex}
\alias{subset_sex.mig_net_rate_age_sex}
\alias{subset_time.mig_parameter}
\alias{subset_indicator.mig_parameter}
\alias{subset_time.mort_rate_age_sex}
\alias{subset_age.mort_rate_age_sex}
\alias{subset_sex.mort_rate_age_sex}
\alias{subset_time.pop_count_age_sex_base}
\alias{subset_age.pop_count_age_sex_base}
\alias{subset_sex.pop_count_age_sex_base}
\alias{subset_time.srb}
\alias{subset_time.survival_ratio_age_sex}
\alias{subset_age.survival_ratio_age_sex}
\alias{subset_sex.survival_ratio_age_sex}
\alias{subset_indicator.ccmpp_output_df}
\alias{subset_time.ccmpp_output_df}
\alias{subset_age.ccmpp_output_df}
\alias{subset_sex.ccmpp_output_df}
\alias{subset_time.pop_count_age_sex}
\alias{subset_age.pop_count_age_sex}
\alias{subset_sex.pop_count_age_sex}
\alias{subset_time.pop_count_age_sex_reference}
\alias{subset_age.pop_count_age_sex_reference}
\alias{subset_sex.pop_count_age_sex_reference}
\alias{subset_demog_change_component_df}
\alias{subset_time}
\alias{subset_time.demog_change_component_df}
\alias{subset_age}
\alias{subset_age.demog_change_component_df}
\alias{subset_sex}
\alias{subset_sex.demog_change_component_df}
\alias{subset_indicator}
\alias{subset_indicator.demog_change_component_df}
\title{Subset by time, age, sex, or indicator}
\usage{
\method{subset_time}{death_count_age_sex}(x, times, include = TRUE)

\method{subset_age}{death_count_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{death_count_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{death_probability_age_sex}(x, times, include = TRUE)

\method{subset_age}{death_probability_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{death_probability_age_sex}(x, sexes, include = TRUE)

\method{subset_indicator}{ccmpp_input_df}(x, indicators, include = TRUE)

\method{subset_time}{ccmpp_input_df}(x, times, include = TRUE)

\method{subset_age}{ccmpp_input_df}(x, ages, include = TRUE)

\method{subset_sex}{ccmpp_input_df}(x, sexes, include = TRUE)

\method{subset_time}{fert_rate_age_f}(x, times, include = TRUE)

\method{subset_age}{fert_rate_age_f}(x, ages, include = TRUE)

\method{subset_indicator}{life_table_age_sex}(x, indicators, include = TRUE)

\method{subset_time}{life_table_age_sex}(x, times, include = TRUE)

\method{subset_age}{life_table_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{life_table_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{mig_net_count_age_sex}(x, times, include = TRUE)

\method{subset_age}{mig_net_count_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{mig_net_count_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{mig_net_count_tot_b}(x, times, include = TRUE)

\method{subset_time}{mig_net_rate_age_sex}(x, times, include = TRUE)

\method{subset_age}{mig_net_rate_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{mig_net_rate_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{mig_parameter}(x, times, include = TRUE)

\method{subset_indicator}{mig_parameter}(x, times, include = TRUE)

\method{subset_time}{mort_rate_age_sex}(x, times, include = TRUE)

\method{subset_age}{mort_rate_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{mort_rate_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{pop_count_age_sex_base}(x, times, include = TRUE)

\method{subset_age}{pop_count_age_sex_base}(x, ages, include = TRUE)

\method{subset_sex}{pop_count_age_sex_base}(x, sexes, include = TRUE)

\method{subset_time}{srb}(x, times, include = TRUE)

\method{subset_time}{survival_ratio_age_sex}(x, times, include = TRUE)

\method{subset_age}{survival_ratio_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{survival_ratio_age_sex}(x, sexes, include = TRUE)

\method{subset_indicator}{ccmpp_output_df}(x, indicators, include = TRUE)

\method{subset_time}{ccmpp_output_df}(x, times, include = TRUE)

\method{subset_age}{ccmpp_output_df}(x, ages, include = TRUE)

\method{subset_sex}{ccmpp_output_df}(x, sexes, include = TRUE)

\method{subset_time}{pop_count_age_sex}(x, times, include = TRUE)

\method{subset_age}{pop_count_age_sex}(x, ages, include = TRUE)

\method{subset_sex}{pop_count_age_sex}(x, sexes, include = TRUE)

\method{subset_time}{pop_count_age_sex_reference}(x, times, include = TRUE)

\method{subset_age}{pop_count_age_sex_reference}(x, ages, include = TRUE)

\method{subset_sex}{pop_count_age_sex_reference}(x, sexes, include = TRUE)

subset_time(x, times, include = TRUE, ...)

\method{subset_time}{demog_change_component_df}(x, times, include = TRUE, drop = FALSE)

subset_age(x, ages, include = TRUE, ...)

\method{subset_age}{demog_change_component_df}(x, ages, include = TRUE, drop = FALSE)

subset_sex(x, sexes, include = TRUE, ...)

\method{subset_sex}{demog_change_component_df}(x, sexes = get_all_allowed_sexes(), include = TRUE, drop = FALSE)

subset_indicator(x, indicators, include = TRUE, ...)

\method{subset_indicator}{demog_change_component_df}(x, indicators, include = TRUE, drop = FALSE)
}
\arguments{
\item{x}{An object to subset.}

\item{include}{Logical; should the rows corresponding to the values
supplied in the previous argument be those that are kept or
discarded?}

\item{indicators, times, ages, sexes}{Vectors indicating the levels of
time, age, or sex to retain (\code{include = TRUE}) or exclude
(\code{include = FALSE}). \code{ages} and \code{times} are
coerced to numeric via \code{\link{as.numeric}}.}

\item{drop}{Logical; should demographic change component dimensions
with only one level be dropped? Not available in all methods;
see \dQuote{Details}.}
}
\value{
The object after subsetting if a valid member of the class;
    otherwise an error.
}
\description{
These functions subset objects inheriting from
\code{demog_change_component_df} by one of the four dimensions,
indicator, age, time, or sex, and retain the class if valid (see
\dQuote{Details}). Subsetting such that only one level of the
subset dimension is retained will drop the dimension if \code{drop
= TRUE}, otherwise it is retained (default). If the object returned
after subsetting is a valid member of the original class it will be
returned (valid according to
\code{\link{validate_ccmppWPP_object}}). If it is not valid an error
will be signalled and nothing is returned.
}
\details{
For \code{subset_times} and \code{subset_ages} rows can be excluded
by supplying negative \code{times} or \code{ages}. In
that case, all values must be negative and \code{include} must be
\code{TRUE} (the default). Violations will result in an error.

Argument \code{drop} is not available in the methods for objects
inheriting from \code{ccmpp_input_df} because these objects must
have (and keep) the set of dimensions specified in their class
definition. E.g., \code{fert_rate_input} objects may be subset by
time, but the time dimension cannot be dropped because objects of
this class must always have a time dimension.
}
\section{Note on efficiency}{

These functions are not particularly efficient. For repeated
subsetting within, e.g., a for loop, it is better to use standard
subsetting operations on the data frame component and re-cast the
result as a classed object at the end if desired.
}

\seealso{
Other subset_by_demographic_dimension: 
\code{\link{subset_ccmpp_input_list}}
}
\author{
Mark Wheldon
}
\concept{subset_by_demographic_dimension}
