% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_validators.R
\name{validate_ccmpp_object}
\alias{validate_ccmpp_object}
\alias{validate_ccmpp_object.default}
\alias{validate_ccmpp_object.demog_change_component_df}
\alias{validate_ccmpp_object.ccmpp_input_df}
\alias{validate_ccmpp_object.fert_rate_input_df}
\alias{validate_ccmpp_object.survival_ratio_input_df}
\alias{validate_ccmpp_object.pop_count_base_input_df}
\alias{validate_ccmpp_object.srb_input_df}
\alias{validate_ccmpp_object.mig_net_rate_input_df}
\alias{validate_ccmpp_object.mig_net_count_input_df}
\alias{validate_ccmpp_object.mig_net_count_tot_input_df}
\alias{validate_ccmpp_object.mig_parameter_input_df}
\alias{validate_ccmpp_object.life_table_input_df}
\alias{validate_ccmpp_object.ccmpp_input_list}
\title{Validate objects of class \code{demog_change_component_df}.}
\usage{
validate_ccmpp_object(x, ...)

\method{validate_ccmpp_object}{default}(x, ...)

\method{validate_ccmpp_object}{demog_change_component_df}(x, ...)

\method{validate_ccmpp_object}{ccmpp_input_df}(x, ...)

\method{validate_ccmpp_object}{fert_rate_input_df}(x, ...)

\method{validate_ccmpp_object}{survival_ratio_input_df}(x, ...)

\method{validate_ccmpp_object}{pop_count_base_input_df}(x, ...)

\method{validate_ccmpp_object}{srb_input_df}(x, ...)

\method{validate_ccmpp_object}{mig_net_rate_input_df}(x, ...)

\method{validate_ccmpp_object}{mig_net_count_input_df}(x, ...)

\method{validate_ccmpp_object}{mig_net_count_tot_input_df}(x, ...)

\method{validate_ccmpp_object}{mig_parameter_input_df}(x, ...)

\method{validate_ccmpp_object}{life_table_input_df}(x, ...)

\method{validate_ccmpp_object}{ccmpp_input_list}(x, ...)
}
\arguments{
\item{x}{An object to be validated.}
}
\value{
Either an error or the object \code{x}.
}
\description{
Checks that an object with \code{class} attribute
\code{demog_change_component_df} is a valid object of this type.
}
\seealso{
\code{\link{demog_change_component_df}} and the other
    creator functions (e.g., \code{\link{ccmpp_input_df}},
    \code{link{fert_rate_input_df}}, etc.)
}
\author{
Mark Wheldon
}
