% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes_demog_change_component_df.R
\name{validate_demog_change_component_df}
\alias{validate_demog_change_component_df}
\title{Validate objects of class \code{demog_change_component_df}.}
\usage{
validate_demog_change_component_df(x)
}
\arguments{
\item{x}{An object of class \code{demog_change_component_df}.}
}
\value{
Either an error or the object \code{x}.
}
\description{
Checks that an object with \code{class} attribute
\code{demog_change_component_df} is a valid object of this type.
}
\seealso{
demog_change_component_df

Other demog_change_component_df class non-exported functions: 
\code{\link{new_demog_change_component_df}()}
}
\author{
Mark Wheldon
}
\concept{demog_change_component_df class non-exported functions}
